Print Summary Statistics of Alpha Level Cutoffs

Print Summary Statistics of Alpha Level Cutoffs

alphaSummary(object, ...)

# S4 method for bcbioRNASeq
alphaSummary(object, alpha = c(0.1, 0.05, 0.01, 0.001,
  1e-06), caption = NULL, ...)

# S4 method for DESeqDataSet
alphaSummary(object, alpha = c(0.1, 0.05, 0.01,
  0.001, 1e-06), caption = NULL, ...)

Arguments

object
Object.

...
Optional. Passthrough arguments to DESeq2::results(). Use either contrast or name arguments to define the desired contrast.

alpha
Numeric vector of desired alpha cutoffs.

caption
Optional. Character vector to add as caption to the table.

Value

kable.

Note

bcbioRNASeq does not support contrast definitions, since the object contains an internal DESeqDataSet with an empty design formula.

See also

DESeq2::results().

Other Differential Expression Utilities: plotDispEsts, plotMeanSD

Examples

load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "dds.rda"),
    package = "bcbioRNASeq"))

# bcbioRNASeq
alphaSummary(bcb)

#> Warning: Empty DESeqDataSet design formula detected

#> 
#> 
#> |               |0.1              |0.05             |0.01             |0.001            |1e-06            |
#> |:--------------|:----------------|:----------------|:----------------|:----------------|:----------------|
#> |LFC > 0 (up)   |170, 56%         |166, 55%         |157, 52%         |152, 50%         |136, 45%         |
#> |LFC < 0 (down) |0, 0%            |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |outliers       |1, 0.33%         |1, 0.33%         |1, 0.33%         |1, 0.33%         |1, 0.33%         |
#> |low counts     |67, 22%          |57, 19%          |0, 0%            |74, 24%          |0, 0%            |
#> |cutoff         |(mean count < 1) |(mean count < 0) |(mean count < 0) |(mean count < 1) |(mean count < 0) |

# DESeqDataSet
alphaSummary(dds)

#> 
#> 
#> |               |0.1              |0.05             |0.01             |0.001            |1e-06            |
#> |:--------------|:----------------|:----------------|:----------------|:----------------|:----------------|
#> |LFC > 0 (up)   |2, 0.66%         |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |LFC < 0 (down) |2, 0.66%         |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |outliers       |0, 0%            |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |low counts     |0, 0%            |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |cutoff         |(mean count < 0) |(mean count < 0) |(mean count < 0) |(mean count < 0) |(mean count < 0) |

alphaSummary(dds, contrast = c("group", "ko", "ctrl"))

#> 
#> 
#> |               |0.1              |0.05             |0.01             |0.001            |1e-06            |
#> |:--------------|:----------------|:----------------|:----------------|:----------------|:----------------|
#> |LFC > 0 (up)   |2, 0.66%         |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |LFC < 0 (down) |2, 0.66%         |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |outliers       |0, 0%            |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |low counts     |0, 0%            |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |cutoff         |(mean count < 0) |(mean count < 0) |(mean count < 0) |(mean count < 0) |(mean count < 0) |

alphaSummary(dds, name = "group_ko_vs_ctrl")

#> 
#> 
#> |               |0.1              |0.05             |0.01             |0.001            |1e-06            |
#> |:--------------|:----------------|:----------------|:----------------|:----------------|:----------------|
#> |LFC > 0 (up)   |2, 0.66%         |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |LFC < 0 (down) |2, 0.66%         |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |outliers       |0, 0%            |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |low counts     |0, 0%            |0, 0%            |0, 0%            |0, 0%            |0, 0%            |
#> |cutoff         |(mean count < 0) |(mean count < 0) |(mean count < 0) |(mean count < 0) |(mean count < 0) |