Force an Object to Belong to a Class

Coerce an object to a given class.

bcbioRNASeq to DESeqDataSet

Coerces to RangedSummarizedExperiment.

Rounds raw counts to integer matrix.

Subsets colData to include only clean factor columns. See sampleData() for more details.

Simplifies metadata to include only relevant information and updates sessionInfo slot.

Note that gene-level counts are required. Alternatively tximport::summarizeToGene() can be called to convert transcript-level counts to gene-level. By default, we're using length-scaled TPM, so a corresponding average transcript length matrix isn't necessary. The average transcript length matrix is only necessary when raw counts matrix isn't scaled during tximport call (see countsFromAbundance in tximport).

bcbioRNASeq to DESeqTransform

Coerces to DESeqDataSet.

Calls DESeq2::DESeq().

Calls DESeq2::varianceStabilizingTransformation().

Description

as(object) returns the version of this object coerced to be the given Class. When used in the replacement form on the left of an assignment, the portion of the object corresponding to Class is replaced by value.

The operation of as() in either form depends on the definition of coerce methods. Methods are defined automatically when the two classes are related by inheritance; that is, when one of the classes is a subclass of the other.

Coerce methods are also predefined for basic classes (including all the types of vectors, functions and a few others).

Beyond these two sources of methods, further methods are defined by calls to the setAs function. See that documentation also for details of how coerce methods work. Use showMethods(coerce) for a list of all currently defined methods, as in the example below.

Basic Coercion Methods

Methods are pre-defined for coercing any object to one of the basic datatypes. For example, as(x, "numeric") uses the existing as.numeric function. These and all other existing methods can be listed as shown in the example.

References

Chambers, John M. (2016) Extending R, Chapman & Hall. (Chapters 9 and 10.)

See also

Examples

data(bcb)

## bcbioRNASeq to DESeqDataSet ====
x <- as(bcb, "DESeqDataSet")

#> Generating DESeqDataSet with DESeq2 1.22.2.

names(S4Vectors::mcols(x))

#> [1] "broadClass"     "description"    "entrezID"       "geneBiotype"   
#> [5] "geneID"         "geneName"       "seqCoordSystem"

class(x)

#> [1] "DESeqDataSet"
#> attr(,"package")
#> [1] "DESeq2"

show(x)

#> class: DESeqDataSet 
#> dim: 500 6 
#> metadata(11): version caller ... date sessionInfo
#> assays(1): counts
#> rownames(500): ENSMUSG00000000088 ENSMUSG00000000326 ...
#>   ENSMUSG00000114367 ENSMUSG00000114458
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(6): control_rep1 control_rep2 ... fa_day7_rep2 fa_day7_rep3
#> colData names(9): qualityFormat sampleName ... tissue treatment

## bcbioRNASeq to RangedSummarizedExperiment ====
x <- as(bcb, "RangedSummarizedExperiment")
slotNames(x)

#> [1] "rowRanges"       "colData"         "assays"          "NAMES"          
#> [5] "elementMetadata" "metadata"       

show(x)

#> class: RangedSummarizedExperiment 
#> dim: 500 6 
#> metadata(28): version level ... subset previousVersion
#> assays(6): counts tpm ... vst fpkm
#> rownames(500): ENSMUSG00000000088 ENSMUSG00000000326 ...
#>   ENSMUSG00000114367 ENSMUSG00000114458
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(6): control_rep1 control_rep2 ... fa_day7_rep2 fa_day7_rep3
#> colData names(25): averageInsertSize averageReadLength ... tissue
#>   treatment

## bcbioRNASeq to SummarizedExperiment ====
## Coerce to RangedSummarizedExperiment first.
x <- as(bcb, "RangedSummarizedExperiment")
x <- as(x, "SummarizedExperiment")
class(x)

#> [1] "SummarizedExperiment"
#> attr(,"package")
#> [1] "SummarizedExperiment"

slotNames(x)

#> [1] "colData"         "assays"          "NAMES"           "elementMetadata"
#> [5] "metadata"       

show(x)

#> class: SummarizedExperiment 
#> dim: 500 6 
#> metadata(28): version level ... subset previousVersion
#> assays(6): counts tpm ... vst fpkm
#> rownames(500): ENSMUSG00000000088 ENSMUSG00000000326 ...
#>   ENSMUSG00000114367 ENSMUSG00000114458
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(6): control_rep1 control_rep2 ... fa_day7_rep2 fa_day7_rep3
#> colData names(25): averageInsertSize averageReadLength ... tissue
#>   treatment