bcbioRNADataSet Legacy Object Class

This class has been superceded by bcbioRNASeq and will be formally deprecated in a future release.

Slots

callers

SimpleList containing additional bcbio run data with dimensions that don't match the count matrix.