bcbioRNASeq Generator

Simply point to the final upload directory generated by bcbio, and this constructor function will take care of the rest. It automatically imports RNA-seq counts, metadata, and the program versions used.

bcbioRNASeq(uploadDir, level = c("genes", "transcripts"),
  caller = c("salmon", "kallisto", "sailfish", "star", "hisat2"),
  samples = NULL, censorSamples = NULL, sampleMetadataFile = NULL,
  organism = NULL, genomeBuild = NULL, ensemblRelease = NULL,
  gffFile = NULL, transgeneNames = NULL, spikeNames = NULL,
  countsFromAbundance = "lengthScaledTPM", vst = TRUE, rlog = FALSE,
  interestingGroups = "sampleName", ...)

Arguments

uploadDir
string. Path to final upload directory. This path is set when running "bcbio_nextgen -w template".

level
string. Import counts at gene level ("genes"; default) or transcript level ("transcripts"; advanced use).

caller
string. Expression caller:

"salmon" (default): Salmon alignment-free, quasi-mapped counts.

"kallisto". Kallisto alignment-free, pseudo-aligned counts.

"sailfish". Sailfish alignment-free, lightweight counts.

"star": STAR (Spliced Transcripts Alignment to a Reference) aligned counts.

"hisat2": HISAT2 (Hierarchical Indexing for Spliced Alignment of Transcripts) graph-based aligned counts.

samples
character or NULL. Specify a subset of samples to load. The names must match the description specified in the bcbio YAML metadata. If a sampleMetadataFile is provided, that will take priority for sample selection. Typically this can be left unset.

censorSamples
character or NULL. Samples to exclude from the analysis.

sampleMetadataFile
string or NULL. Custom metadata file containing sample information. Otherwise defaults to sample metadata saved in the YAML file. Remote URLs are supported. Typically this can be left unset.

organism
string or NULL. Organism name. Use the full Latin name (e.g. "Homo sapiens"), since this will be input downstream to AnnotationHub and ensembldb, unless gffFile is set. If left NULL (not recommended), the function call will skip loading gene/transcript-level annotations into rowRanges(). This can be useful for poorly annotation genomes or experiments involving multiple genomes.

genomeBuild
string or NULL. Ensembl genome build name (e.g. "GRCh38"). This will be passed to AnnotationHub for EnsDb annotation matching, unless gffFile is set.

ensemblRelease
scalar integer or NULL. Ensembl release version. If unset, defaults to current release, and does not typically need to be user-defined. Passed to AnnotationHub for EnsDb annotation matching, unless gffFile is set.

gffFile
string or NULL. By default, we recommend leaving this NULL for genomes that are supported on Ensembl. In this case, the row annotations (rowRanges()) will be obtained automatically from Ensembl by passing the organism, genomeBuild, and ensemblRelease arguments to AnnotationHub and ensembldb. For a genome that is not supported on Ensembl and/or AnnotationHub, a GFF/GTF (General Feature Format) file is required. Generally, we recommend using a GTF (GFFv2) file here over a GFF3 file if possible, although all GFF formats are supported. The function will internally generate a TxDb containing transcript-to-gene mappings and construct a GRanges object containing the genomic ranges (rowRanges()).

transgeneNames
character or NULL. Vector indicating which assay() rows denote transgenes (e.g. EGFP, TDTOMATO).

spikeNames
character or NULL. Vector indicating which assay() rows denote spike-in sequences (e.g. ERCCs).

countsFromAbundance
character, either "no" (default), "scaledTPM", "lengthScaledTPM", or "dtuScaledTPM". Whether to generate estimated counts using abundance estimates:

scaled up to library size (scaledTPM),

scaled using the average transcript length over samples and then the library size (lengthScaledTPM), or

scaled using the median transcript length among isoforms of a gene, and then the library size (dtuScaledTPM).

dtuScaledTPM is designed for DTU analysis in combination with txOut=TRUE, and it requires specifing a tx2gene data.frame. dtuScaledTPM works such that within a gene, values from all samples and all transcripts get scaled by the same fixed median transcript length. If using scaledTPM, lengthScaledTPM, or geneLengthScaledTPM, the counts are no longer correlated across samples with transcript length, and so the length offset matrix should not be used.

vst
boolean. Calculate variance-stabilizing transformation using DESeq2::varianceStabilizingTransformation(). Recommended by default for visualization.

rlog
boolean. Calcualte regularized log transformation using DESeq2::rlog(). This calculation is slow for large datasets and now discouraged by default for visualization.

interestingGroups
character. Groups of interest that define the samples. If left unset, defaults to sampleName.

...
Additional arguments.

Value

bcbioRNASeq.

Details

This class contains raw read counts and length-scaled transcripts per million (TPM) generated by tximport::tximport(). Counts can be loaded at gene or transcript level.

Metadata

The metadata() slot contains:

Sample quality control metrics.

Ensembl annotations.

Server run paths.

R session information (e.g. utils::sessionInfo()).

Valid names

R is strict about values that are considered valid for use in names() and dimnames() (i.e. rownames() and colnames()). Non-alphanumeric characters, spaces, and dashes are not valid. Use either underscores or periods in place of dashes when working in R. Also note that names should not begin with a number, and will be prefixed with an X when sanitized. Consult the documentation in the base::make.names() function for more information. We strongly recommend adhering to these conventions when labeling samples, to help avoid unexpected downstream behavior in R due to dimnames() mismatches.

Genome build

Ensure that the organism and genome build used with bcio match correctly here in the function call. In particular, for the legacy Homo sapiens GRCh37/hg19 genome build, ensure that genomeBuild = "GRCh37". Otherwise, the genomic ranges set in rowRanges() will mismatch. It is recommended for current projects that GRCh38/hg38 is used in place of GRCh37/hg19 if possible.

DESeq2

DESeq2 is run automatically when bcbioRNASeq() is called. Internally, this automatically slots normalized counts into assays(), and optionally generates variance-stabilized rlog or vst counts, depending on the call. When loading a dataset with a large number of samples (i.e. > 50), we recommend disabling the rlog transformation, since it can take a long time to compute.

Remote data

When working in RStudio, we recommend connecting to the bcbio run directory as a remote connection over sshfs. When loading a large number of samples, it is preferable to call bcbioRNASeq() directly in R on the remote server, if possible.

See also

Examples

uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")

# Gene level
object <- bcbioRNASeq(
    uploadDir = uploadDir,
    level = "genes",
    caller = "salmon",
    organism = "Mus musculus",
    ensemblRelease = 87L
)

#> Dated project directory: 2017-05-23_rnaseq

#> 4 sample(s) detected:
#> group1_1, group1_2, group2_1, group2_2

#> Reading project-summary.yaml...

#> Reading sample metadata from YAML...

#> Reading project-summary.yaml...

#> Reading sample metrics from YAML...

#> Reading project-summary.yaml...

#> Parsed with column specification:
#> cols(
#>   transcriptID = col_character(),
#>   geneID = col_character()
#> )

#> Reading salmon counts using tximport 1.9.12

#> 1 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 2 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 3 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 4 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 

#> summarizing abundance

#> summarizing counts

#> summarizing length

#> Generating DESeqDataSet with DESeq2 1.21.22...

#> using just counts from tximport

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

#> Applying variance-stabilizing transformation...

#> Using `makeGRangesFromEnsembl()` for annotations.

#> Making GRanges from Ensembl with AnnotationHub 2.13.7 (2018-10-04)...

#> Ensembl 87 EnsDb for Mus Musculus

#>   - AnnotationHub: "AH53222"
#>   - Organism: "Mus musculus"
#>   - Build: "GRCm38"
#>   - Release: "87"
#>   - Level: "genes"

#> Defining broadClass using: geneName, geneBiotype, seqnames

#> Arranging by geneID...

#> Warning: 'DESCRIPTION' file has an 'Encoding' field and re-encoding is not possible

print(object)

#> bcbioRNASeq 0.99.0
#> http://bioinformatics.sph.harvard.edu/bcbioRNASeq
#> citation("bcbioRNASeq")
#> ────────────────────────────────────────────────────────────────────────
#> Upload Dir: "/home/bioinfo/git/bcbioRNASeq-private/inst/extdata/bcbio"
#> Upload Date: 2017-05-23
#> R Load Date: 2018-10-09
#> Level: "genes"
#> Caller: "salmon"
#> Organism: "Mus musculus"
#> Interesting Groups: "sampleName"
#> ────────────────────────────────────────────────────────────────────────
#> class: RangedSummarizedExperiment 
#> dim: 502 4 
#> metadata(26): version level ... wd sessionInfo
#> assays(4): counts tpm normalized vst
#> rownames(502): ENSMUSG00000002459 ENSMUSG00000004768 ...
#>   ENSMUSG00000104523 ENSMUSG00000105982
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(4): group1_1 group1_2 group2_1 group2_2
#> colData names(21): averageInsertSize duplicates ... group samRef

# Transcript level
object <- bcbioRNASeq(
    uploadDir = uploadDir,
    level = "transcripts",
    caller = "salmon",
    organism = "Mus musculus",
    ensemblRelease = 87L
)

#> Dated project directory: 2017-05-23_rnaseq

#> 4 sample(s) detected:
#> group1_1, group1_2, group2_1, group2_2

#> Reading project-summary.yaml...

#> Reading sample metadata from YAML...

#> Reading project-summary.yaml...

#> Reading sample metrics from YAML...

#> Reading project-summary.yaml...

#> Parsed with column specification:
#> cols(
#>   transcriptID = col_character(),
#>   geneID = col_character()
#> )

#> Reading salmon counts using tximport 1.9.12

#> 1 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 2 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 3 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 4 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 

#> Using `makeGRangesFromEnsembl()` for annotations.

#> Making GRanges from Ensembl with AnnotationHub 2.13.7 (2018-10-04)...

#> Ensembl 87 EnsDb for Mus Musculus

#>   - AnnotationHub: "AH53222"
#>   - Organism: "Mus musculus"
#>   - Build: "GRCm38"
#>   - Release: "87"
#>   - Level: "transcripts"

#> Defining broadClass using: geneName, transcriptBiotype, seqnames

#> Arranging by transcriptID...

#> Warning: 'DESCRIPTION' file has an 'Encoding' field and re-encoding is not possible

print(object)

#> bcbioRNASeq 0.99.0
#> http://bioinformatics.sph.harvard.edu/bcbioRNASeq
#> citation("bcbioRNASeq")
#> ────────────────────────────────────────────────────────────────────────
#> Upload Dir: "/home/bioinfo/git/bcbioRNASeq-private/inst/extdata/bcbio"
#> Upload Date: 2017-05-23
#> R Load Date: 2018-10-09
#> Level: "transcripts"
#> Caller: "salmon"
#> Organism: "Mus musculus"
#> Interesting Groups: "sampleName"
#> ────────────────────────────────────────────────────────────────────────
#> class: RangedSummarizedExperiment 
#> dim: 996 4 
#> metadata(26): version level ... wd sessionInfo
#> assays(2): counts tpm
#> rownames(996): ENSMUST00000002533 ENSMUST00000006037 ...
#>   ENSMUST00000208793 ENSMUST00000209020
#> rowData names(13): broadClass description ... transcriptName
#>   transcriptSupportLevel
#> colnames(4): group1_1 group1_2 group2_1 group2_2
#> colData names(21): averageInsertSize duplicates ... group samRef