bcbioRNASeq Class
Source: R/AllClasses.R

S4 class containing RNA-seq counts and metadata generated by bcbio.

Details

bcbioRNASeq extends RangedSummarizedExperiment and is designed to store a bcbio RNA-seq analysis. This class contains raw read counts and length-scaled transcripts per million (TPM) generated by tximport::tximport(). Counts can be loaded at gene (default) or transcript level.

DESeq2 is run automatically when loadRNASeq() is called, and variance stabilized counts are slotted into assays().

The metadata() accessor contains:

Sample quality control metrics.

Ensembl annotations.

Server run paths.

R session information (e.g. utils::sessionInfo()).

Note

bcbioRNASeq extended SummarizedExperiment prior to v0.2.0, where we migrated to RangedSummarizedExperiment.

See also

Examples

show(bcb_small)

#> bcbioRNASeq 0.2.0
#> samples: 6
#> genes: 500
#> organism: Mus musculus
#> bcbio dir: /mnt/O2/bcbio/bcbioRNASeq/F1000v2/GSE65267-merged/final
#> bcbio date: 2018-02-18
#> load date: 2018-03-13

validObject(bcb_small)

#> [1] TRUE