Export

Export data out of R and write to disk.

# S4 method for DESeqResultsTables
export(x, dir = ".", compress = TRUE,
  dropbox = FALSE, rdsToken = NULL)

Arguments

x
Object.

dir
string. Directory path.

compress
boolean. Apply gzip compression to all files.

dropbox
boolean. Export results to Dropbox instead of local storage. Recommended method by HBC for permanent storage (e.g. Stem Cell Commons). When this option is enabled, unique links per file are generated internally with bcbioBase::copyToDropbox(), which relies on the rdrop2 package. Note that local files are written to base::tempdir() when this option is enabled.

rdsToken
string or NULL. RDS file token to use for Dropbox authentication. If set NULL and dropbox = TRUE then an interactive prompt will appear requesting authorization.

Value

Invisible character. File path(s).

Details

This is a wrapper for rio::export() that adds support for additional S4 classes in Bioconductor.

Row names

The standard rio::export() call will drop rownames when exporting to CSV. The readr::write_*() family of functions also never write rownames. This is a really poor default setting for handling genomic data, which often contain gene identifiers in the rownames. Here we're performing any internal tibble coercion step to ensure rownames are always moved to a "rowname" column in the CSV export.

See also

Examples

data(deseq_small)
x <- DESeqResultsTables(deseq_small)

#> DESeqResults: treatment folic acid vs control (shrunken LFC)

#> Adding `rowData()` annotations (atomic columns only).
#> [1] "broadClass"     "description"    "geneBiotype"    "geneID"        
#> [5] "geneName"       "seqCoordSystem"

export(x, dir = "example")

#> Writing treatment_folic_acid_vs_control_all.csv.gz, treatment_folic_acid_vs_control_up.csv.gz, treatment_folic_acid_vs_control_down.csv.gz, treatment_folic_acid_vs_control_both.csv.gz to /Users/mike/git/bcbioRNASeq-private/docs/reference/example.

#> Exported treatment_folic_acid_vs_control_all.csv.gz.

#> Exported treatment_folic_acid_vs_control_up.csv.gz.

#> Exported treatment_folic_acid_vs_control_down.csv.gz.

#> Exported treatment_folic_acid_vs_control_both.csv.gz.

#> DESeqResultsTables 1.21.22
#> http://bioinformatics.sph.harvard.edu/bcbioRNASeq
#> citation("bcbioRNASeq")
#> contrast: treatment folic acid vs control
#> alpha: 0.1
#> lfcThreshold: 0
#> dir: /Users/mike/git/bcbioRNASeq-private/docs/reference/example
#> out of 500 with nonzero total read count
#> adjusted p-value < 0.1
#> LFC > 0 (up)       : 115, 23%
#> LFC < 0 (down)     : 139, 28%
#> outliers [1]       : 6, 1.2%
#> low counts [2]     : 0, 0%
#> (mean count < 33)

list.files("example")

#> [1] "treatment_folic_acid_vs_control_all.csv.gz" 
#> [2] "treatment_folic_acid_vs_control_both.csv.gz"
#> [3] "treatment_folic_acid_vs_control_down.csv.gz"
#> [4] "treatment_folic_acid_vs_control_up.csv.gz"  

## Clean up.
unlink("example", recursive = TRUE)