Extract or Replace Parts of an Object

Extract genes by row and samples by column from a bcbioRNASeq object. Internal count transformations are rescaled automatically, if defined.

# S4 method for bcbioRNASeq,ANY,ANY,ANY
[(x, i, j, drop = FALSE,
  transform = TRUE)

Arguments

x
object from which to extract element(s) or in which to replace element(s).

i
indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object (or for matrices/arrays, the dimnames): see ‘Character indices’ below for further details.

For [-indexing only: i, j, … can be logical vectors, indicating elements/slices to select. Such vectors are recycled if necessary to match the corresponding extent. i, j, … can also be negative integers, indicating elements/slices to leave out of the selection.

When indexing arrays by [ a single argument i can be a matrix with as many columns as there are dimensions of x; the result is then a vector with elements corresponding to the sets of indices in each row of i.

An index value of NULL is treated as if it were integer(0).

j
indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object (or for matrices/arrays, the dimnames): see ‘Character indices’ below for further details.

For [-indexing only: i, j, … can be logical vectors, indicating elements/slices to select. Such vectors are recycled if necessary to match the corresponding extent. i, j, … can also be negative integers, indicating elements/slices to leave out of the selection.

When indexing arrays by [ a single argument i can be a matrix with as many columns as there are dimensions of x; the result is then a vector with elements corresponding to the sets of indices in each row of i.

An index value of NULL is treated as if it were integer(0).

drop
For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension (see the examples). This only works for extracting elements, not for the replacement. See drop for further details.

transform
boolean. Recalculate slotted DESeq2 transformations. Recommended by default but can take a while for large datasets.

Value

bcbioRNASeq.

Note

DESeq2 transformations will only be calculated when transform = TRUE and either rlog or vst counts are defined in assays().

See also

Examples

object <- bcb_small

# Minimum of 100 genes, 2 samples
genes <- head(rownames(object), 100L)
head(genes)

#> [1] "ENSMUSG00000000088" "ENSMUSG00000000326" "ENSMUSG00000000346"
#> [4] "ENSMUSG00000000740" "ENSMUSG00000001175" "ENSMUSG00000001666"

samples <- head(colnames(object), 2L)
head(samples)

#> [1] "control_rep1" "control_rep2"

# Extract by sample name
object[, samples]

#> Updating normalized counts

#> Generating DESeqDataSet with DESeq2 1.20.0

#> using just counts from tximport

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

#> Recalculating DESeq2 variance stabilizations (transform = TRUE)

#> Applying variance-stabilizing transformation

#> bcbioRNASeq 0.2.7
#> class: RangedSummarizedExperiment 
#> dim: 500 2 
#> metadata(31): version level ... subset originalVersion
#> assays(5): counts tpm length normalized vst
#> rownames(500): ENSMUSG00000000088 ENSMUSG00000000326 ...
#>   ENSMUSG00000114367 ENSMUSG00000114458
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(2): control_rep1 control_rep2
#> colData names(26): sampleName day ... x5x3Bias xGC
#> ================================================================================
#> Upload Dir: "/mnt/O2/bcbio/bcbioRNASeq/F1000v2/GSE65267-merged/final"
#> Upload Date: 2018-03-18
#> R Load Date: 2018-08-23
#> Level: "genes"
#> Caller: "salmon"
#> Organism: "Mus musculus"
#> Interesting Groups: "treatment"
#> AnnotationHub: "AH57770"
#> Ensembl Release: "90"
#> Genome Build: "GRCm38"

# Extract by gene list
object[genes, ]

#> Updating normalized counts

#> Generating DESeqDataSet with DESeq2 1.20.0

#> using just counts from tximport

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

#> Recalculating DESeq2 variance stabilizations (transform = TRUE)

#> Applying variance-stabilizing transformation

#> bcbioRNASeq 0.2.7
#> class: RangedSummarizedExperiment 
#> dim: 100 6 
#> metadata(31): version level ... subset originalVersion
#> assays(5): counts tpm length normalized vst
#> rownames(100): ENSMUSG00000000088 ENSMUSG00000000326 ...
#>   ENSMUSG00000024503 ENSMUSG00000024610
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(6): control_rep1 control_rep2 ... fa_day7_rep2 fa_day7_rep3
#> colData names(26): sampleName day ... x5x3Bias xGC
#> ================================================================================
#> Upload Dir: "/mnt/O2/bcbio/bcbioRNASeq/F1000v2/GSE65267-merged/final"
#> Upload Date: 2018-03-18
#> R Load Date: 2018-08-23
#> Level: "genes"
#> Caller: "salmon"
#> Organism: "Mus musculus"
#> Interesting Groups: "treatment"
#> AnnotationHub: "AH57770"
#> Ensembl Release: "90"
#> Genome Build: "GRCm38"

# Extract by both genes and samples
x <- object[genes, samples]

#> Updating normalized counts

#> Generating DESeqDataSet with DESeq2 1.20.0

#> using just counts from tximport

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

#> Recalculating DESeq2 variance stabilizations (transform = TRUE)

#> Applying variance-stabilizing transformation

print(x)

#> bcbioRNASeq 0.2.7
#> class: RangedSummarizedExperiment 
#> dim: 100 2 
#> metadata(31): version level ... subset originalVersion
#> assays(5): counts tpm length normalized vst
#> rownames(100): ENSMUSG00000000088 ENSMUSG00000000326 ...
#>   ENSMUSG00000024503 ENSMUSG00000024610
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(2): control_rep1 control_rep2
#> colData names(26): sampleName day ... x5x3Bias xGC
#> ================================================================================
#> Upload Dir: "/mnt/O2/bcbio/bcbioRNASeq/F1000v2/GSE65267-merged/final"
#> Upload Date: 2018-03-18
#> R Load Date: 2018-08-23
#> Level: "genes"
#> Caller: "salmon"
#> Organism: "Mus musculus"
#> Interesting Groups: "treatment"
#> AnnotationHub: "AH57770"
#> Ensembl Release: "90"
#> Genome Build: "GRCm38"

assayNames(x)

#> [1] "counts"     "tpm"        "length"     "normalized" "vst"       

# Fast subsetting, by skipping DESeq2 transformations
# Note that rlog and vst counts will not be defined
x <- object[, samples, transform = FALSE]

#> Updating normalized counts

#> Generating DESeqDataSet with DESeq2 1.20.0

#> using just counts from tximport

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

#> Skipping DESeq2 variance stabilizations (transform = FALSE)

print(x)

#> bcbioRNASeq 0.2.7
#> class: RangedSummarizedExperiment 
#> dim: 500 2 
#> metadata(31): version level ... subset originalVersion
#> assays(4): counts tpm length normalized
#> rownames(500): ENSMUSG00000000088 ENSMUSG00000000326 ...
#>   ENSMUSG00000114367 ENSMUSG00000114458
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(2): control_rep1 control_rep2
#> colData names(26): sampleName day ... x5x3Bias xGC
#> ================================================================================
#> Upload Dir: "/mnt/O2/bcbio/bcbioRNASeq/F1000v2/GSE65267-merged/final"
#> Upload Date: 2018-03-18
#> R Load Date: 2018-08-23
#> Level: "genes"
#> Caller: "salmon"
#> Organism: "Mus musculus"
#> Interesting Groups: "treatment"
#> AnnotationHub: "AH57770"
#> Ensembl Release: "90"
#> Genome Build: "GRCm38"

names(assays(x))

#> [1] "counts"     "tpm"        "length"     "normalized"