Heatmaps

These functions are wrappers to pheatmap() that have been optimized to work with a bcbioRnaDataSet run object. They also depend on a DESeqTransform object containing transformed counts, which are ideal for heatmap plots.

plot_correlation_heatmap(run, dt, genes = NULL, intgroup = NULL,
  annotation = NULL, title = NULL, method = "pearson",
  clustering_method = "ward.D2", ...)

plot_gene_heatmap(run, dt, genes, intgroup = NULL, annotation = NULL,
  title = NULL, clustering_method = "ward.D2", cluster_rows = TRUE,
  cluster_cols = TRUE, scale = "row", ...)

plot_deg_heatmap(run, dt, res, ...)

Arguments

run
bcbioRnaDataSet.

dt
DESeqTransform. Transformed counts generated by rlog() (recommended) or vst() on a DESeqDataSet are supported.

genes
Character vector of specific gene identifiers to plot.

intgroup
Interesting groups to label with bars above heatmap. If NULL, defaults to intgroup defined in bcbioRnaDataSet.

annotation
Optional alternative annotation to use. Useful when plotting more than one column.

title
Optional text to include in plot title.

method
Correlation coefficient (or covariance) to be computed. Defaults to pearson but spearman can also be used.

clustering_method
Clustering method. Accepts the same values as hclust().

...
Passthrough parameters to pheatmap().

cluster_rows
Whether rows should be clustered using hclust().

cluster_cols
Whether columns should be clustered using hclust().

scale
Character indicating if the values should be centered and scaled in either the row direction, column direction, or none.

res
DESeqResults.

Value

pheatmap() object.

Details

plot_correlation_heatmap() calculates a correlation matrix based on gene expression per sample. By default, this function processes all gene counts per sample to calculate the corrlation matrix. This behavior can be overrided with the input of gene identifier vector. In this case, only the expression of the desired genes will be used to calculate the correlation matrix. plot_gene_heatmap() facilities heatmap plotting of a specified set of genes. The function by default performs row- and column-wise hierarhical clustering using the Ward method, but this behavior can be overrided by setting cluster_rows or cluster_cols to FALSE. When column clustering is disabled, the function sorts columns by the interesting groups (intgroup) specified in the bcbioRnaDataSet and then the sample names. plot_deg_heatmap() is a simplified version of plot_gene_heatmap() that is optimized for handling a DESeqResults object rather a gene vector. All of the optional parameters for plot_gene_heatmap() are also available to this function.

See also

cor().

pheatmap().