Load bcbio RNA-Seq Data

Simply point to the final upload directory output by bcbio, and this function will take care of the rest. It automatically imports RNA-seq counts, metadata, and the program versions used.

loadRNASeq(uploadDir, level = c("genes", "transcripts"),
  caller = c("salmon", "kallisto", "sailfish"), samples = NULL,
  sampleMetadataFile = NULL, interestingGroups = "sampleName", organism,
  ensemblRelease = NULL, genomeBuild = NULL, isSpike = NULL,
  gffFile = NULL, transformationLimit = 50L, ...)

Arguments

uploadDir
Path to final upload directory. This path is set when running bcbio_nextgen -w template.

level
Import counts as "genes" (default) or "transcripts".

caller
Expression caller. Supports "salmon" (default), "kallisto", or "sailfish".

samples
Optional. Specify a subset of samples to load. The names must match the description specified in the bcbio YAML metadata. If a sampleMetadataFile is provided, that will take priority for sample selection. Typically this can be left unset.

sampleMetadataFile
Optional. Custom metadata file containing sample information. Otherwise defaults to sample metadata saved in the YAML file. Remote URLs are supported. Typically this can be left unset.

interestingGroups
Character vector denoting groups of interest that define the samples. If left unset, defaults to sampleName.

organism
Organism name. Use the full latin name (e.g. "Homo sapiens"), since this will be input downstream to AnnotationHub and ensembldb, unless gffFile is set.

ensemblRelease
Optional. Ensembl release version. If unset, defaults to current release, and does not typically need to be user-defined. Passed to AnnotationHub for EnsDb annotation matching, unless gffFile is set.

genomeBuild
Optional. Ensembl genome build name (e.g. "GRCh38"). This will be passed to AnnotationHub for EnsDb annotation matching, unless gffFile is set.

isSpike
Optional. Gene names corresponding to FASTA spike-in sequences (e.g. ERCCs, EGFP, TDTOMATO).

gffFile
Optional, not recommended. By default, we recommend leaving this NULL for genomes that are supported on Ensembl. In this case, the row annotations (rowRanges()) will be obtained automatically from Ensembl by passing the organism, genomeBuild, and ensemblRelease arguments to AnnotationHub and ensembldb. For a genome that is not supported on Ensembl and/or AnnotationHub, a GFF/GTF (General Feature Format) file is required. Generally, we recommend using a GTF (GFFv2) file here over a GFF3 file if possible, although all GFF formats are supported. The function will internally generate a TxDb containing transcript-to-gene mappings and construct a GRanges object containing the genomic ranges (rowRanges()).

transformationLimit
Maximum number of samples to calculate DESeq2::rlog() and DESeq2::varianceStabilizingTransformation() matrix. For large datasets, DESeq2 is slow to apply variance stabilization. In this case, we recommend loading up the dataset in a high-performance computing environment. Use Inf to always apply and -Inf to always skip.

...
Additional arguments, slotted into the metadata() accessor.

Value

bcbioRNASeq.

Details

When the number of samples is bigger than the transformationLimit, rlog and vst counts will not be slotted into assays(). In this case, we recommend visualization using tmm() counts, which are automatically calculated using edgeR.

Note

When working in RStudio, we recommend connecting to the bcbio-nextgen run directory as a remote connection over sshfs.

See also

help("bcbioRNASeq-class", "bcbioRNASeq").

Examples

uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")

# Gene level
loadRNASeq(
    uploadDir = uploadDir,
    level = "genes",
    organism = "Mus musculus"
)

#> 2017-05-23_rnaseq

#> 4 samples detected:
#> group1_1, group1_2, group2_1, group2_2

#> Reading project-summary.yaml

#> Fetching Ensembl genes from AnnotationHub 2.10.1 (2017-10-27)

#> hub id: "AH57770"
#> organism: "Mus musculus"
#> genome build: "GRCm38"
#> release version: "90"

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Defining broadClass using: geneName, geneBiotype, seqnames

#> Reading sample metrics

#> Obtaining transcript-to-gene mappings

#> Parsed with column specification:
#> cols(
#>   txID = col_character(),
#>   geneID = col_character()
#> )

#> Reading salmon counts using tximport

#> 1 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 2 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 3 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 4 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 

#> summarizing abundance

#> summarizing counts

#> summarizing length

#> Calculating variance stabilizations using DESeq2 1.18.1

#> using just counts from tximport

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

#> Applying rlog transformation

#> Applying variance stabilizing transformation

#> Warning: Dropping 3 unannotated rows (0.594%): ENSMUSG00000101738, ENSMUSG00000104475, ENSMUSG00000109048

#> bcbioRNASeq 0.2.0
#> samples: 4
#> genes: 502
#> organism: Mus musculus
#> bcbio dir: /Users/mike/GitHub/hbc/bcbioRNASeq/inst/extdata/bcbio
#> bcbio date: 2017-05-23
#> load date: 2018-03-22

# Transcript level
loadRNASeq(
    uploadDir = uploadDir,
    level = "transcripts",
    organism = "Mus musculus"
)

#> 2017-05-23_rnaseq

#> 4 samples detected:
#> group1_1, group1_2, group2_1, group2_2

#> Reading project-summary.yaml

#> Fetching Ensembl transcripts from AnnotationHub 2.10.1 (2017-10-27)

#> hub id: "AH57770"
#> organism: "Mus musculus"
#> genome build: "GRCm38"
#> release version: "90"

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Setting txBiotype as factor

#> Setting txSupportLevel as factor

#> Defining broadClass using: geneName, txBiotype, seqnames

#> Reading sample metrics

#> Obtaining transcript-to-gene mappings

#> Parsed with column specification:
#> cols(
#>   txID = col_character(),
#>   geneID = col_character()
#> )

#> Reading salmon counts using tximport

#> 1 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 2 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 3 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 4 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 

#> Warning: Dropping 1 unannotated rows (0.1%): ENSMUST00000187966

#> bcbioRNASeq 0.2.0
#> samples: 4
#> transcripts: 998
#> organism: Mus musculus
#> bcbio dir: /Users/mike/GitHub/hbc/bcbioRNASeq/inst/extdata/bcbio
#> bcbio date: 2017-05-23
#> load date: 2018-03-22