Load bcbio-nextgen run

Simply point to the final upload directory output by bcbio-nextgen, and this function will take care of the rest. It automatically imports RNA-seq counts, metadata, and program versions used.

load_run(upload_dir = "final", analysis = "rnaseq",
  interesting_groups = "description", design = NULL, contrast = NULL,
  sample_metadata_file = NULL)

Arguments

upload_dir
Path to final upload directory. This path is set when running bcbio_nextgen -w template.

analysis
Analysis type. Supports RNA-seq (rnaseq; default) or small RNA-seq (srnaseq).

interesting_groups
Character vector of interesting groups. First entry is used for plot colors during quality control (QC) analysis. Entire vector is used for PCA and heatmap QC functions.

sample_metadata_file
(Optional). Custom metadata file containing sample information. Otherwise defaults to sample metadata saved in the YAML file.

experiment_name
Experiment name.

principal_investigator
Principal investigator.

researcher
Researcher who performed the experiment.

email
Email for follow-up correspondence.

Value

bcbioRNADataSet.

Note

When working in RStudio, we recommend connecting to the bcbio-nextgen run directory as a remote connection over sshfs.