Melt Count Matrix to Long Format and log10 Transform

Melt Count Matrix to Long Format and log10 Transform

meltLog10(object, ...)

# S4 method for bcbioRNADataSet
meltLog10(object, normalized = TRUE)

# S4 method for DESeqDataSet
meltLog10(object, normalized = TRUE,
  interestingGroups = NULL)

# S4 method for DESeqTransform
meltLog10(object, interestingGroups = NULL)

Arguments

object
Object.

...
Additional arguments (for the S4 generic definition).

normalized
Select normalized counts (TRUE), raw counts (FALSE), or specifically request TMM-normalized counts (tmm).

interestingGroups
Optional. Interesting groups.

Value

log10 melted data.frame.

See also

reshape2::melt().

Examples

data(bcb, dds, rld)

# bcbioRNADataSet
meltLog10(bcb) %>% glimpse

#> Observations: 922
#> Variables: 5
#> $ ensgene      <chr> "ENSMUSG00000002459", "ENSMUSG00000004768", "ENSMUSG00...
#> $ sampleID     <chr> "group1_1", "group1_1", "group1_1", "group1_1", "group...
#> $ counts       <dbl> 1.12058610, 2.64346485, 3.27516222, 2.87263455, 0.0414...
#> $ sampleName   <chr> "group1_1", "group1_1", "group1_1", "group1_1", "group...
#> $ sampleName.1 <chr> "group1_1", "group1_1", "group1_1", "group1_1", "group...

# DESeqDataSet
meltLog10(dds) %>% glimpse

#> Observations: 922
#> Variables: 6
#> $ ensgene    <chr> "ENSMUSG00000002459", "ENSMUSG00000004768", "ENSMUSG0000...
#> $ sampleID   <chr> "group1_1", "group1_1", "group1_1", "group1_1", "group1_...
#> $ counts     <dbl> 1.12058610, 2.64346485, 3.27516222, 2.87263455, 0.041404...
#> $ sampleName <chr> "group1_1", "group1_1", "group1_1", "group1_1", "group1_...
#> $ group      <fctr> ctrl, ctrl, ctrl, ctrl, ctrl, ctrl, ctrl, ctrl, ctrl, c...
#> $ sizeFactor <dbl> 0.9090654, 0.9090654, 0.9090654, 0.9090654, 0.9090654, 0...

# DESeqTransform
meltLog10(rld) %>% glimpse

#> Observations: 841
#> Variables: 7
#> $ ensgene    <chr> "ENSMUSG00000002459", "ENSMUSG00000004768", "ENSMUSG0000...
#> $ sampleID   <chr> "group1_1", "group1_1", "group1_1", "group1_1", "group1_...
#> $ counts     <dbl> 0.5335118, 0.9380281, 1.0354722, 0.9718929, 0.7577204, 0...
#> $ colname    <chr> "group1_1", "group1_1", "group1_1", "group1_1", "group1_...
#> $ sampleName <chr> "group1_1", "group1_1", "group1_1", "group1_1", "group1_...
#> $ group      <fctr> ctrl, ctrl, ctrl, ctrl, ctrl, ctrl, ctrl, ctrl, ctrl, c...
#> $ sizeFactor <dbl> 0.9090654, 0.9090654, 0.9090654, 0.9090654, 0.9090654, 0...