Plot sexually dimorphic gender marker genes

This is a convenience function that quickly plots known sexually dimorphic genes, and overall gene experssion for the X and Y chromosomes.

plotGenderMarkers(object, ...)

# S4 method for bcbioRNASeq
plotGenderMarkers(object, normalized = c("vst",
  "rlog", "tmm", "rle", "tpm", "fpkm"), interestingGroups = NULL,
  trans = c("identity", "log2", "log10"), medianLine = TRUE,
  color = getOption(x = "acid.color.discrete", default =
  acidplots::scale_colour_synesthesia_d()), legend = getOption(x =
  "acid.legend", default = TRUE), style = "wide")

Arguments

object
Object.

normalized
character(1). Type of normalized counts to visualize:

tpm Transcripts per million.

tmm: edgeR trimmed mean of M-values.

rlog: DESeq2 log2 regularized log transformation.

vst: DESeq2 log2 variance stabilizing transformation.

interestingGroups
character. Groups of interest to use for visualization. Corresponds to factors describing the columns of the object.

trans
character(1). Name of the axis scale transformation to apply.

For more information:

help(topic = "scale_x_continuous", package = "ggplot2")

medianLine
logical(1). Include median line for each group. Disabled if samples are colored by sample name.

color
ScaleDiscrete. Desired ggplot2 color scale. Must supply discrete values. When set NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_color_manual().

To set the discrete color palette globally, use:

options(acid.color.discrete = ggplot2::scale_color_viridis_d())

legend
logical(1). Include plot legend.

style
character(1). Plot style.

...
Additional arguments.

Value

ggplot.

Note

Updated 2019-08-07.

Methods (by class)

SingleCellExperiment: Applies pseudobulk() calculation to average gene expression at sample level prior to plotting.

See also

Examples

data(bcb)
plotGenderMarkers(bcb)

#> Error in (function (object, genes, strict = TRUE) {    validObject(object)    assert(is(object, "Gene2Symbol"), isCharacter(genes), isFlag(strict))    if (any(genes %in% rownames(object))) {        table <- rownames(object)    }    else if (any(genes %in% object[["geneName"]])) {        assert(matchesUniqueGeneNames(object, genes))        table <- object[["geneName"]]    }    else if (any(genes %in% object[["geneID"]])) {        table <- object[["geneID"]]    }    else {        stop(sprintf("All genes failed to map: %s.", toString(genes,             width = 100L)))    }    match <- match(x = genes, table = table)    names(match) <- genes    if (isTRUE(strict)) {        fun <- stop    }    else {        fun <- message    }    unmapped <- which(is.na(match))    if (length(unmapped) > 0L) {        fun(sprintf("Some genes failed to map: %s.", toString(genes[unmapped],             width = 200L)))    }    mapped <- na.omit(match)    assert(hasLength(mapped))    mapped})(object = new("Gene2Symbol", rownames = c("ENSMUSG00000000001", "ENSMUSG00000000003", "ENSMUSG00000000028", "ENSMUSG00000000049", "ENSMUSG00000000058", "ENSMUSG00000000078", "ENSMUSG00000000085", "ENSMUSG00000000088", "ENSMUSG00000000093", "ENSMUSG00000000094", "ENSMUSG00000000120", "ENSMUSG00000000125", "ENSMUSG00000000126", "ENSMUSG00000000127", "ENSMUSG00000000149", "ENSMUSG00000000157", "ENSMUSG00000000159", "ENSMUSG00000000167", "ENSMUSG00000000171", "ENSMUSG00000000182", "ENSMUSG00000000183", "ENSMUSG00000000184", "ENSMUSG00000000197", "ENSMUSG00000000202", "ENSMUSG00000000204", "ENSMUSG00000000214", "ENSMUSG00000000215", "ENSMUSG00000000216", "ENSMUSG00000000223", "ENSMUSG00000000247", "ENSMUSG00000000248", "ENSMUSG00000000253", "ENSMUSG00000000275", "ENSMUSG00000000276", "ENSMUSG00000000278", "ENSMUSG00000000282", "ENSMUSG00000000290", "ENSMUSG00000000295", "ENSMUSG00000000296", "ENSMUSG00000000301", "ENSMUSG00000000303", "ENSMUSG00000000305", "ENSMUSG00000000308", "ENSMUSG00000000317", "ENSMUSG00000000318", "ENSMUSG00000000320", "ENSMUSG00000000326", "ENSMUSG00000000339", "ENSMUSG00000000340", "ENSMUSG00000000346", "ENSMUSG00000000355", "ENSMUSG00000000359", "ENSMUSG00000000374", "ENSMUSG00000000378", "ENSMUSG00000000384", "ENSMUSG00000000385", "ENSMUSG00000000392", "ENSMUSG00000000402", "ENSMUSG00000000411", "ENSMUSG00000000420", "ENSMUSG00000000435", "ENSMUSG00000000439", "ENSMUSG00000000440", "ENSMUSG00000000441", "ENSMUSG00000000489", "ENSMUSG00000000530", "ENSMUSG00000000531", "ENSMUSG00000000532", "ENSMUSG00000000560", "ENSMUSG00000000562", "ENSMUSG00000000567", "ENSMUSG00000000581", "ENSMUSG00000000594", "ENSMUSG00000000605", "ENSMUSG00000000617", "ENSMUSG00000000627", "ENSMUSG00000000628", "ENSMUSG00000000631", "ENSMUSG00000000632", "ENSMUSG00000000673", "ENSMUSG00000020152", "ENSMUSG00000020333", "ENSMUSG00000020676", "ENSMUSG00000020875", "ENSMUSG00000026315", "ENSMUSG00000026535", "ENSMUSG00000028494", "ENSMUSG00000029231", "ENSMUSG00000029534", "ENSMUSG00000029730", "ENSMUSG00000032050", "ENSMUSG00000035352", "ENSMUSG00000035385", "ENSMUSG00000038257", "ENSMUSG00000044807", "ENSMUSG00000048583", "ENSMUSG00000055022", "ENSMUSG00000061689", "ENSMUSG00000062661", "ENSMUSG00000074340"), nrows = 100L, listData = list(geneID = c("ENSMUSG00000000001", "ENSMUSG00000000003", "ENSMUSG00000000028", "ENSMUSG00000000049", "ENSMUSG00000000058", "ENSMUSG00000000078", "ENSMUSG00000000085", "ENSMUSG00000000088", "ENSMUSG00000000093", "ENSMUSG00000000094", "ENSMUSG00000000120", "ENSMUSG00000000125", "ENSMUSG00000000126", "ENSMUSG00000000127", "ENSMUSG00000000149", "ENSMUSG00000000157", "ENSMUSG00000000159", "ENSMUSG00000000167", "ENSMUSG00000000171", "ENSMUSG00000000182", "ENSMUSG00000000183", "ENSMUSG00000000184", "ENSMUSG00000000197", "ENSMUSG00000000202", "ENSMUSG00000000204", "ENSMUSG00000000214", "ENSMUSG00000000215", "ENSMUSG00000000216", "ENSMUSG00000000223", "ENSMUSG00000000247", "ENSMUSG00000000248", "ENSMUSG00000000253", "ENSMUSG00000000275", "ENSMUSG00000000276", "ENSMUSG00000000278", "ENSMUSG00000000282", "ENSMUSG00000000290", "ENSMUSG00000000295", "ENSMUSG00000000296", "ENSMUSG00000000301", "ENSMUSG00000000303", "ENSMUSG00000000305", "ENSMUSG00000000308", "ENSMUSG00000000317", "ENSMUSG00000000318", "ENSMUSG00000000320", "ENSMUSG00000000326", "ENSMUSG00000000339", "ENSMUSG00000000340", "ENSMUSG00000000346", "ENSMUSG00000000355", "ENSMUSG00000000359", "ENSMUSG00000000374", "ENSMUSG00000000378", "ENSMUSG00000000384", "ENSMUSG00000000385", "ENSMUSG00000000392", "ENSMUSG00000000402", "ENSMUSG00000000411", "ENSMUSG00000000420", "ENSMUSG00000000435", "ENSMUSG00000000439", "ENSMUSG00000000440", "ENSMUSG00000000441", "ENSMUSG00000000489", "ENSMUSG00000000530", "ENSMUSG00000000531", "ENSMUSG00000000532", "ENSMUSG00000000560", "ENSMUSG00000000562", "ENSMUSG00000000567", "ENSMUSG00000000581", "ENSMUSG00000000594", "ENSMUSG00000000605", "ENSMUSG00000000617", "ENSMUSG00000000627", "ENSMUSG00000000628", "ENSMUSG00000000631", "ENSMUSG00000000632", "ENSMUSG00000000673", "ENSMUSG00000020152", "ENSMUSG00000020333", "ENSMUSG00000020676", "ENSMUSG00000020875", "ENSMUSG00000026315", "ENSMUSG00000026535", "ENSMUSG00000028494", "ENSMUSG00000029231", "ENSMUSG00000029534", "ENSMUSG00000029730", "ENSMUSG00000032050", "ENSMUSG00000035352", "ENSMUSG00000035385", "ENSMUSG00000038257", "ENSMUSG00000044807", "ENSMUSG00000048583", "ENSMUSG00000055022", "ENSMUSG00000061689", "ENSMUSG00000062661", "ENSMUSG00000074340"), geneName = c("Gnai3", "Pbsn", "Cdc45", "Apoh", "Cav2", "Klf6", "Scmh1", "Cox5a", "Tbx2", "Tbx4", "Ngfr", "Wnt3", "Wnt9a", "Fer", "Gna12", "Itgb2l", "Igsf5", "Pih1d2", "Sdhd", "Fgf23", "Fgf6", "Ccnd2", "Nalcn", "Btbd17", "Slfn4", "Th", "Ins2", "Scnn1g", "Drp2", "Lhx2", "Clec2g", "Gmpr", "Trim25", "Dgke", "Scpep1", "Mnt", "Itgb2", "Hddc2", "Tpd52l1", "Pemt", "Cdh1", "Cdh4", "Ckmt1", "Bcl6b", "Clec10a", "Alox12", "Comt", "Rtca", "Dbt", "Dazap2", "Mcts1", "Rem1", "Trappc10", "Ccm2", "Tbrg4", "Tmprss2", "Fap", "Egfl6", "Tssk3", "Galnt1", "Myf5", "Mkrn2", "Pparg", "Raf1", "Pdgfb", "Acvrl1", "Grasp", "Acvr1b", "Gabra2", "Adora3", "Sox9", "C1d", "Gm2a", "Clcn4", "Grm6", "Sema4f", "Hk2", "Myo18a", "Sez6", "Haao", "Actr2", "Acsl6", "Ccl11", "Hoxb9", "Serpinb8", "Ifi202b", "Plin2", "Pdgfra", "St7", "Mcm7", "Rdx", "Ccl12", "Ccl2", "Glra3", "Zfp354c", "Igf2", "Cntn1", "Dlgap4", "Ncs1", "Ovgp1")), elementType = "ANY",     elementMetadata = NULL, metadata = list(package = "freerange",         version = structure(list(c(0L, 2L, 0L)), class = c("package_version",         "numeric_version")), date = structure(18099, class = "Date"),         organism = "Mus musculus", genomeBuild = "GRCm38", ensemblRelease = 90L,         ensembldb = structure(list(name = c("Db type", "Type of Gene ID",         "Supporting package", "Db created by", "script_version",         "Creation time", "ensembl_version", "ensembl_host", "Organism",         "taxonomy_id", "genome_build", "DBSCHEMAVERSION"), value = c("EnsDb",         "Ensembl Gene ID", "ensembldb", "ensembldb package from Bioconductor",         "0.3.4", "Tue Jan  1 21:15:02 2019", "90", "localhost",         "Mus musculus", "10090", "GRCm38", "2.1")), class = "data.frame", row.names = c(NA,         -12L)), level = "genes", id = "AH57770", format = "makeUnique")),     genes = c("ENSMUSG00000086503", "ENSMUSG00000069045", "ENSMUSG00000056673",     "ENSMUSG00000068457"), strict = FALSE): All genes failed to map: ENSMUSG00000086503, ENSMUSG00000069045, ENSMUSG00000056673, ENSMUSG00000068457.