Gene heatmap

These functions facilitate heatmap plotting of a specified set of genes. By default, row- and column-wise hierarchical clustering is performed using the Ward method, but this behavior can be overrided by setting cluster_rows or cluster_cols to FALSE. When column clustering is disabled, the columns are sorted by the interesting groups (interesting_groups) specified in the bcbioRNADataSet and then the sample names.

plot_gene_heatmap(bcb, dt = NULL, genes, clustering_method = "ward.D2",
  cluster_rows = TRUE, cluster_cols = TRUE, scale = "row",
  interesting_groups = NULL, annotation = NULL, title = NULL, ...)

plot_deg_heatmap(bcb, dt, res, lfc, title = NULL, ...)

Arguments

bcb
bcbioRNADataSet.

dt
DESeqTransform. Transformed counts generated by rlog() (recommended) or vst() on a DESeqDataSet are supported.

genes
Character vector of specific gene identifiers to plot.

clustering_method
Hierarchical clustering method. Accepts the same values as stats::hclust().

cluster_rows
Use hierarchical clustering to arrange rows.

cluster_cols
Use hierarchical clustering to arrange columns.

scale
Character indicating if the values should be centered and scaled in either the row direction, column direction, or none.

interesting_groups
Optional. Interesting groups to label with bars above heatmap. If NULL, defaults to interesting_groups defined in the bcbioRNADataSet.

annotation
Optional. Alternative annotation to use. Useful when plotting more than one column.

title
Optional. Text to include in plot title.

...
Additional arguments.

res
DESeqResults.

lfc
log2 fold change ratio.

Details

plot_deg_heatmap() is a simplified version of plot_gene_heatmap() that is optimized for handling a DESeqResults object rather a gene vector. All of the optional parameters for plot_gene_heatmap() are also available to this function.

See also

pheatmap::pheatmap().