Quality control plots

Plot count density

Counts per gene plot

Wrapper for DESeq2::plotPCA() that improves PCA sample coloring and labeling

Exonic mapping rate plot

Plot sexually dimorphic gender marker genes

Genes detection saturation plot

Genes detected plot

Intronic mapping rate plot

Mapped reads plot

Mapping rate plot

rRNA contamination mapping rate

Plot total reads

plot_count_density(bcbio, counts)

plot_counts_per_gene(bcbio, normalized_counts)

plot_deseq_pca(bcbio, dt, label = TRUE)

plot_exonic_mapping_rate(bcbio)

plot_gender_markers(bcbio, normalized_counts)

plot_gene_detection_saturation(bcbio, raw_counts)

plot_genes_detected(bcbio, raw_counts)

plot_intronic_mapping_rate(bcbio)

plot_mapped_reads(bcbio)

plot_mapping_rate(bcbio)

plot_rrna_mapping_rate(bcbio)

plot_total_reads(bcbio)

Arguments

bcbio

bcbio-nextgen run object

normalized_counts

Normalized counts matrix. Can be obtained from DESeqDataSet by running counts(normalized = TRUE). Transcripts per million (TPM) are also acceptable.

dt

DESeqTransform generated from rlog() or vst() on a DESeqDataSet object. We prefer to pipe the transformed data in to this function because the operation is CPU intensive, and is slow when applied to multiple PCA plot operations.

label

Superimpose sample text labels on the plot

raw_counts

Raw counts matrix. Can be obtained from DESeqDataSet by running counts(normalized = FALSE).

Value

Density plot

Boxplot

PCA plot

Bar plot

Scatter plot

Smooth plot

Bar plot

Bar plot

Bar plot

Bar plot

Bar plot

Bar plot