Read data from bcbio-nextgen run

Read RNA-seq counts using tximport::tximport(). Currently supports salmon (preferred) and sailfish.

Read a project data file.

Read sample metadata from YAML.

Read summary metrics from YAML. These statistics are only generated for a standard RNA-seq run with aligned counts. Fast RNA-seq mode with lightweight counts (pseudocounts) doesn't output the same metrics into the YAML.

Read bcbio sample information from YAML.

read_bcbio_counts(run, grep = NULL, samples = NULL)

read_bcbio_file(run, file, row_names = NULL, ...)

read_bcbio_metadata(run)

read_bcbio_metrics(run)

read_bcbio_samples_yaml(run, ...)

Arguments

run
bcbio-nextgen run.

grep
Use grep pattern to match samples. This will override the samples parameter if set.

samples
Character vector of sample names in bcbio final directory to include. If NULL (default), all samples will be imported.

file
File name.

row_names
Column identifier to use for row names.

...
Passthrough parameters for readr importers (readr::read_csv(), readr::read_tsv()).

...
Nested operator keys supplied as dot objects.

Value

Counts saved in txi list object.

Miscellaneous data frame.

Metadata data frame.

Summary statistics data frame.

Sample information data frame.

See also

tximport::tximport().

readr::read_csv().

readr::read_tsv().