Differential Expression Results Tables

Differential Expression Results Tables

resultsTables(results, counts, ...)

# S4 method for DESeqResults,DESeqDataSet
resultsTables(results, counts, alpha,
  lfcThreshold = 0L, summary = TRUE, write = FALSE, headerLevel = 2L,
  dir = ".", dropboxDir = NULL, rdsToken = NULL)

Arguments

results
DESeqResults object.

counts
DESeqDataSet containing the counts that were used to generate the DESeqResults.

...
Additional arguments.

alpha
Adjusted P value ("alpha") cutoff.

lfcThreshold
Log fold change ratio (base 2) cutoff threshold.

summary
Show summary statistics.

write
Write CSV files to disk.

headerLevel
R Markdown header level.

dir
Local directory path.

dropboxDir
Dropbox directory path where to archive the results tables for permanent storage (e.g. Stem Cell Commons). When this option is enabled, unique links per file are generated internally with the rdrop2 package.

rdsToken
RDS file token to use for Dropbox authentication.

Value

list containing modified DESeqResults return, including additional gene-level metadata and normalized counts.

Note

Log fold change cutoff threshold ("lfcThreshold") does not apply to statistical hypothesis testing, only gene filtering in the results tables. See DESeq2::results() for additional information about using lfcThreshold and altHypothesis to set an alternative hypothesis based on expected fold changes.

See also

Examples

# DESeqResults, DESeqDataSet ====
x <- resultsTables(
    results = res_small,
    counts = dds_small,
    lfcThreshold = 0.25
)

#> 
#> 
#> ## Summary statistics
#> 
#> 
#> - 500 genes in counts matrix
#> - Base mean > 0: 500 genes (non-zero)
#> - Base mean > 1: 500 genes
#> - Alpha: 0.1
#> - LFC threshold: 0.25
#> - DEG pass alpha: 254 genes
#> - DEG LFC up: 115 genes
#> - DEG LFC down: 139 genes
#> 

names(x)

#> [1] "deg"          "degLFC"       "degLFCUp"     "degLFCDown"   "all"         
#> [6] "contrast"     "alpha"        "lfcThreshold"

glimpse(x[["deg"]])

#> Observations: 254
#> Variables: 23
#> $ geneID         <chr> "ENSMUSG00000024164", "ENSMUSG00000029304", "ENSMUSG...
#> $ baseMean       <dbl> 12562.2432, 69339.5105, 3560.1421, 3544.1246, 3001.7...
#> $ log2FoldChange <dbl> 8.271529, 5.355205, 4.079595, 3.564108, 3.742423, 5....
#> $ lfcSE          <dbl> 0.2736384, 0.2861386, 0.2828642, 0.2553612, 0.284346...
#> $ stat           <dbl> 29.931566, 18.716632, 14.412404, 13.950754, 13.15278...
#> $ pvalue         <dbl> 7.645607e-197, 3.623490e-78, 4.323962e-47, 3.112976e...
#> $ padj           <dbl> 3.776930e-194, 8.950019e-76, 7.120125e-45, 3.844525e...
#> $ seqnames       <fct> 17, 5, 3, 14, 11, 2, 10, 6, 17, 18, 7, 11, 17, 2, 16...
#> $ start          <int> 57203970, 104435118, 19957054, 65968483, 46735080, 3...
#> $ end            <int> 57228136, 104441050, 20009145, 65981547, 46779578, 3...
#> $ width          <int> 24167, 5933, 52092, 13065, 44499, 3620, 4991, 3389, ...
#> $ strand         <fct> -, +, +, +, +, -, -, -, +, +, -, -, -, -, +, +, +, -...
#> $ geneName       <chr> "C3", "Spp1", "Cp", "Clu", "Havcr1", "Lcn2", "Lyz2",...
#> $ geneBiotype    <fct> protein_coding, protein_coding, protein_coding, prot...
#> $ description    <chr> "complement component 3 [Source:MGI Symbol;Acc:MGI:8...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...
#> $ broadClass     <fct> coding, coding, coding, coding, coding, coding, codi...
#> $ controlRep1    <dbl> 7.813753e+01, 2.313604e+03, 3.015620e+02, 5.127776e+...
#> $ controlRep2    <dbl> 46.84188, 2998.40658, 344.73518, 536.83951, 452.1030...
#> $ controlRep3    <dbl> 5.622846e+01, 2.733252e+03, 3.798850e+02, 4.457134e+...
#> $ faDay7Rep1     <dbl> 25465.2621, 182255.3222, 8797.0724, 7941.0384, 4726....
#> $ faDay7Rep2     <dbl> 25909.906, 121258.268, 6227.052, 5552.369, 5293.052,...
#> $ faDay7Rep3     <dbl> 23817.0835, 104478.2110, 5310.5459, 6276.0101, 6902....