Sample Data

Return the sample metadata. Columns are always sanitized to factor.

# S4 method for bcbioRNASeq
sampleData(object, interestingGroups,
  clean = FALSE)

# S4 method for bcbioRNASeq,DataFrame
sampleData(object) <- value

Arguments

object
Object.

interestingGroups
character. Groups of interest that define the samples. If left unset, defaults to sampleName.

clean
boolean. Only return factor columns not defined in bcbioBase::metadataBlacklist.

value
Value to assign.

Value

DataFrame.

Note

This is a complement to the standard colData() function, but improves support for accessing sample metadata for datasets where multiple items in the columns map to a single sample (e.g. cells for a single-cell RNA-seq experiment).

See also

Examples

# SummarizedExperiment ====
sampleData(bcb_small, clean = TRUE) %>% glimpse()

#> Formal class 'DataFrame' [package "S4Vectors"] with 6 slots
#>   ..@ rownames       : chr [1:6] "control_rep1" "control_rep2" "control_rep3" "fa_day7_rep1" ...
#>   ..@ nrows          : int 6
#>   ..@ listData       :List of 6
#>   .. ..$ sampleName: Factor w/ 6 levels "control_rep1",..: 1 2 3 4 5 6
#>   .. ..$ day       : Factor w/ 2 levels "0","7": 1 1 1 2 2 2
#>   .. ..$ replicate : Factor w/ 3 levels "1","2","3": 1 2 3 1 2 3
#>   .. ..$ strain    : Factor w/ 1 level "Balb/c": 1 1 1 1 1 1
#>   .. ..$ tissue    : Factor w/ 1 level "kidney": 1 1 1 1 1 1
#>   .. ..$ treatment : Factor w/ 2 levels "control","folic_acid": 1 1 1 2 2 2
#>   ..@ elementType    : chr "ANY"
#>   ..@ elementMetadata: NULL
#>   ..@ metadata       : list()

sampleData(bcb_small, clean = FALSE) %>% glimpse()

#> Formal class 'DataFrame' [package "S4Vectors"] with 6 slots
#>   ..@ rownames       : chr [1:6] "control_rep1" "control_rep2" "control_rep3" "fa_day7_rep1" ...
#>   ..@ nrows          : int 6
#>   ..@ listData       :List of 27
#>   .. ..$ sampleName                   : Factor w/ 6 levels "control_rep1",..: 1 2 3 4 5 6
#>   .. ..$ day                          : Factor w/ 2 levels "0","7": 1 1 1 2 2 2
#>   .. ..$ description                  : Factor w/ 6 levels "control_rep1",..: 1 2 3 4 5 6
#>   .. ..$ genomeBuild                  : Factor w/ 1 level "GRCm38_90": 1 1 1 1 1 1
#>   .. ..$ replicate                    : Factor w/ 3 levels "1","2","3": 1 2 3 1 2 3
#>   .. ..$ samRef                       : Factor w/ 1 level "/n/app/bcbio/dev/genomes/Mmusculus/GRCm38_90/seq/GRCm38_90.fa": 1 1 1 1 1 1
#>   .. ..$ strain                       : Factor w/ 1 level "Balb/c": 1 1 1 1 1 1
#>   .. ..$ tissue                       : Factor w/ 1 level "kidney": 1 1 1 1 1 1
#>   .. ..$ treatment                    : Factor w/ 2 levels "control","folic_acid": 1 1 1 2 2 2
#>   .. ..$ averageInsertSize            : num [1:6] 243 254 238 235 248 ...
#>   .. ..$ averageReadLength            : num [1:6] 50 50 50 50 50 50
#>   .. ..$ duplicates                   : num [1:6] 0 0 0 0 0 0
#>   .. ..$ duplicationRateOfMapped      : num [1:6] 0.539 0.542 0.566 0.453 0.442 ...
#>   .. ..$ exonicRate                   : num [1:6] 0.734 0.718 0.787 0.655 0.646 ...
#>   .. ..$ intergenicRate               : num [1:6] 0.0634 0.0655 0.0461 0.0487 0.0586 0.0702
#>   .. ..$ intronicRate                 : num [1:6] 0.203 0.217 0.167 0.296 0.296 ...
#>   .. ..$ mappedPairedReads            : num [1:6] 5.94e+07 1.23e+08 5.32e+07 6.79e+07 5.94e+07 ...
#>   .. ..$ mappedReads                  : num [1:6] 5.94e+07 1.23e+08 5.32e+07 6.79e+07 5.94e+07 ...
#>   .. ..$ qualityFormat                : Factor w/ 1 level "standard": 1 1 1 1 1 1
#>   .. ..$ rrna                         : num [1:6] 16412535 31779476 15967974 14933148 13229240 ...
#>   .. ..$ rrnaRate                     : num [1:6] 0.614 0.585 0.645 0.555 0.558 ...
#>   .. ..$ sequenceLength               : num [1:6] 50 50 50 50 50 50
#>   .. ..$ sequencesFlaggedAsPoorQuality: num [1:6] 0 0 0 0 0 0
#>   .. ..$ totalReads                   : num [1:6] 9.72e+07 1.79e+08 7.68e+07 8.11e+07 7.41e+07 ...
#>   .. ..$ x5x3Bias                     : num [1:6] 0.88 0.77 0.91 0.45 0.76 0.65
#>   .. ..$ xGC                          : num [1:6] 46 45 43 40 41 46
#>   .. ..$ interestingGroups            : Factor w/ 2 levels "control","folic_acid": 1 1 1 2 2 2
#>   .. .. ..- attr(*, "names")= chr [1:6] "control_rep1" "control_rep2" "control_rep3" "fa_day7_rep1" ...
#>   ..@ elementType    : chr "ANY"
#>   ..@ elementMetadata: NULL
#>   ..@ metadata       : list()

# Assignment support
x <- bcb_small
sampleData(x)[["test"]] <- seq_len(ncol(x))
# `test` column should be now defined
"test" %in% colnames(sampleData(x))

#> [1] TRUE