Show an Object

Display the object, by printing, plotting or whatever suits its class. This function exists to be specialized by methods. The default method calls showDefault.

Formal methods for show will usually be invoked for automatic printing (see the details).

# S4 method for bcbioRNASeq
show(object)

Arguments

object
Any R object

Value

show returns an invisible NULL.

Details

Objects from an S4 class (a class defined by a call to setClass) will be displayed automatically is if by a call to show. S4 objects that occur as attributes of S3 objects will also be displayed in this form; conversely, S3 objects encountered as slots in S4 objects will be printed using the S3 convention, as if by a call to print.

Methods defined for show will only be inherited by simple inheritance, since otherwise the method would not receive the complete, original object, with misleading results. See the simpleInheritanceOnly argument to setGeneric and the discussion in setIs for the general concept.

See also

Examples

show(bcb_small)

#> bcbioRNASeq 0.2.6
#> class: RangedSummarizedExperiment 
#> dim: 500 6 
#> metadata(30): version level ... devtoolsSessionInfo subset
#> assays(5): counts tpm length normalized vst
#> rownames(500): ENSMUSG00000000088 ENSMUSG00000000326 ...
#>   ENSMUSG00000114367 ENSMUSG00000114458
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(6): control_rep1 control_rep2 ... fa_day7_rep2 fa_day7_rep3
#> colData names(26): sampleName day ... x5x3Bias xGC
#> Upload Dir: "/mnt/O2/bcbio/bcbioRNASeq/F1000v2/GSE65267-merged/final"
#> Upload Date: 2018-03-18
#> R Load Date: 2018-08-19
#> Level: "genes"
#> Caller: "salmon"
#> Organism: "Mus musculus"
#> Interesting Groups: "treatment"
#> AnnotationHub: "AH57770"
#> Ensembl Release: "90"
#> Genome Build: "GRCm38"