Slot aligned counts into object

This function loads aligned counts (e.g. STAR, HISAT2) from the bcbio final output directory into the bcbioRNASeq object, so we can visually inspect correlations with the primary pseudoaligned counts.

slotAlignedCounts(object, ...)

# S4 method for bcbioRNASeq
slotAlignedCounts(object)

Arguments

object
Object.

...
Additional arguments.

Note

Updated 2019-08-07.

Examples

uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")
## Fast mode skips import of aligned counts.
bcb <- bcbioRNASeq(uploadDir, fast = TRUE)

#> Dated project directory: 2018-03-18_GSE65267-merged.

#> 6 samples detected:
#> [1] "control_rep1" "control_rep2" "control_rep3" "fa_day7_rep1" "fa_day7_rep2"
#> [6] "fa_day7_rep3"

#> Importing 'project-summary.yaml' using 'yaml::yaml.load_file()'.

#> Data versions are missing.

#> Importing 'programs.txt' using 'data.table::fread()'.

#> Setting row names from 'rowname' column.

#> Importing 'bcbio-nextgen.log' using 'base::readLines()'.

#> 'bcbio-nextgen.log' file is empty.

#> Importing 'bcbio-nextgen-commands.log' using 'base::readLines()'.

#> 'bcbio-nextgen-commands.log' file is empty.

#> Importing 'tx2gene.csv' using 'data.table::fread()'.

#> Setting row names from 'rowname' column.

#> Getting sample metadata from YAML.

#> Getting sample metrics from YAML.

#> Reading salmon transcript-level counts from quant.sf files using tximport 1.12.3.

#> Scaling transcripts using lengthScaledTPM.

#> Returning transcript abundance at gene level.

#> reading in files with read_tsv

#> 1 

#> 2 

#> 3 

#> 4 

#> 5 

#> 6 

#> 

#> summarizing abundance

#> summarizing counts

#> summarizing length

#> bcbio GTF file: /n/app/bcbio/dev/genomes/Mmusculus/GRCm38_90/rnaseq/ref-transcripts.gtf.

#> bcbio GTF file is not accessible.

#> Slotting empty ranges into 'rowRanges()'.

"aligned" %in% assayNames(bcb)

#> [1] FALSE

bcb <- slotAlignedCounts(bcb)

#> Importing aligned counts from featureCounts.

#> Importing 'combined.counts' using 'data.table::fread()'.

"aligned" %in% assayNames(bcb)

#> [1] TRUE