Trimmed Mean of M-Values (TMM) Normalization

TMM normalization is recommended for RNA-seq data generally when the majority of genes are not differentially expressed. We use this as a quality control tool when plotting counts per gene.

tmm(object)

# S4 method for bcbioRNASeq
tmm(object)

# S4 method for DESeqDataSet
tmm(object)

# S4 method for matrix
tmm(object)

Arguments

object
Object.

Value

matrix.

See also

Other Data Functions: colData, counts, rowData, selectSamples, subset, tpm

Examples

load(system.file("extdata/bcb_small.rda", package = "bcbioRNASeq"))
load(system.file("extdata/dds_small.rda", package = "bcbioRNASeq"))

# bcbioRNASeq ====
tmm(bcb_small) %>% summary()

#> Applying trimmed mean of M-values (TMM) normalization

#>   control_rep1       control_rep2       control_rep3       fa_day7_rep1      
#>  Min.   :     0.0   Min.   :     0.0   Min.   :     0.0   Min.   :     7.73  
#>  1st Qu.:   166.7   1st Qu.:   174.9   1st Qu.:   160.7   1st Qu.:   205.85  
#>  Median :   405.8   Median :   387.7   Median :   390.4   Median :   410.44  
#>  Mean   :  2373.5   Mean   :  2217.7   Mean   :  2352.3   Mean   :  1747.00  
#>  3rd Qu.:  1058.2   3rd Qu.:  1136.7   3rd Qu.:  1029.2   3rd Qu.:  1068.13  
#>  Max.   :200089.3   Max.   :173113.1   Max.   :181335.9   Max.   :110247.98  
#>   fa_day7_rep2       fa_day7_rep3      
#>  Min.   :     0.0   Min.   :     0.29  
#>  1st Qu.:   202.7   1st Qu.:   207.71  
#>  Median :   430.2   Median :   420.84  
#>  Mean   :  1750.2   Mean   :  1690.46  
#>  3rd Qu.:  1062.8   3rd Qu.:   983.82  
#>  Max.   :163481.1   Max.   :161546.00  

# DESeqDataSet ====
tmm(dds_small) %>% summary()

#> Applying trimmed mean of M-values (TMM) normalization

#>   control_rep1       control_rep2       control_rep3       fa_day7_rep1      
#>  Min.   :     0.0   Min.   :     0.0   Min.   :     0.0   Min.   :     7.71  
#>  1st Qu.:   166.5   1st Qu.:   174.9   1st Qu.:   160.7   1st Qu.:   205.88  
#>  Median :   405.6   Median :   387.7   Median :   390.5   Median :   410.47  
#>  Mean   :  2371.9   Mean   :  2217.9   Mean   :  2353.1   Mean   :  1747.14  
#>  3rd Qu.:  1057.5   3rd Qu.:  1136.8   3rd Qu.:  1029.5   3rd Qu.:  1068.18  
#>  Max.   :199957.8   Max.   :173127.9   Max.   :181395.7   Max.   :110256.35  
#>   fa_day7_rep2       fa_day7_rep3      
#>  Min.   :     0.0   Min.   :     0.37  
#>  1st Qu.:   202.8   1st Qu.:   207.70  
#>  Median :   430.2   Median :   420.91  
#>  Mean   :  1750.4   Mean   :  1690.59  
#>  3rd Qu.:  1062.9   3rd Qu.:   983.95  
#>  Max.   :163494.8   Max.   :161558.41  

# matrix ====
counts(bcb_small) %>% tmm() %>% summary()

#> Applying trimmed mean of M-values (TMM) normalization

#>   control_rep1       control_rep2       control_rep3       fa_day7_rep1      
#>  Min.   :     0.0   Min.   :     0.0   Min.   :     0.0   Min.   :     7.73  
#>  1st Qu.:   166.7   1st Qu.:   174.9   1st Qu.:   160.7   1st Qu.:   205.85  
#>  Median :   405.8   Median :   387.7   Median :   390.4   Median :   410.44  
#>  Mean   :  2373.5   Mean   :  2217.7   Mean   :  2352.3   Mean   :  1747.00  
#>  3rd Qu.:  1058.2   3rd Qu.:  1136.7   3rd Qu.:  1029.2   3rd Qu.:  1068.13  
#>  Max.   :200089.3   Max.   :173113.1   Max.   :181335.9   Max.   :110247.98  
#>   fa_day7_rep2       fa_day7_rep3      
#>  Min.   :     0.0   Min.   :     0.29  
#>  1st Qu.:   202.7   1st Qu.:   207.71  
#>  Median :   430.2   Median :   420.84  
#>  Mean   :  1750.2   Mean   :  1690.46  
#>  3rd Qu.:  1062.8   3rd Qu.:   983.82  
#>  Max.   :163481.1   Max.   :161546.00