Update an object to its current class definition

updateObject is a generic function that returns an instance of object updated to its current class definition.

# S4 method for bcbioRNASeq
updateObject(object, rowRanges = NULL)

Arguments

object
Object to be updated for updateObject and updateObjectFromSlots.

Object for slot information to be extracted from for getObjectSlots.

rowRanges
GRanges or NULL. Row annotations. Since we converted to RangedSummarizedExperiment in v0.2.0, this option had to be added to enable updating of newly required rowRanges slot. Objects that are >= v0.2 don't require this argument and it can be left NULL.

Value

bcbioRNASeq.

Details

Update old objects created by the bcbioRNASeq package. The session information metadata is preserved from the time when the bcbio data was originally loaded into R.

Note

Updated 2019-08-12.

Legacy bcbioRNADataSet class

Support for bcbioRNADataSet objects was dropped in v0.2.0 of the package. If you need to load one of these objects, please install an older release.

Legacy bcbioRnaseq package

The previous bcbioRnaseq package (note case) must be reinstalled to load objects from versions <= 0.0.20. We changed the name of the package to bcbioRNASeq starting in v0.0.21.

See also

Examples

data(bcb)
updateObject(bcb)

#> Updating bcbioRNASeq object from version 0.3.22 to 0.3.27.

#> bcbioRNASeq 0.3.27
#> uploadDir: /data00/draco/acidbase/packages/bcbioRNASeq/inst/extdata/bcbio
#> dates(2): [bcbio] 2018-03-18; [R] 2019-07-23
#> level: genes
#> caller: salmon
#> organism: Mus musculus
#> interestingGroups(2): treatment day
#> class: RangedSummarizedExperiment 
#> dim: 100 6 
#> metadata(27): version level ... sessionInfo previousVersion
#> assays(7): counts tpm ... vst fpkm
#> rownames(100): ENSMUSG00000000001 ENSMUSG00000000003 ...
#>   ENSMUSG00000062661 ENSMUSG00000074340
#> rowData names(8): broadClass description ... geneName seqCoordSystem
#> colnames(6): control_rep1 control_rep2 ... fa_day7_rep2 fa_day7_rep3
#> colData names(25): averageInsertSize averageReadLength ... tissue
#>   treatment

## Example that depends on remote file.
## > x <- import(file.path(bcbioRNASeqTestsURL, "bcbioRNASeq_0.1.4.rds"))
## > x <- updateObject(x)
## > x