bcbioRNASeq Object and Constructor

bcbioRNASeq is an S4 class that extends RangedSummarizedExperiment, and is designed to store a bcbio RNA-seq analysis.

bcbioRNASeq(uploadDir, level = c("genes", "transcripts"),
  caller = c("salmon", "kallisto", "sailfish"), organism = NULL,
  sampleMetadataFile = NULL, samples = NULL, censorSamples = NULL,
  interestingGroups = "sampleName", ensemblRelease = NULL,
  genomeBuild = NULL, transgeneNames = NULL, spikeNames = NULL,
  gffFile = NULL, transformationLimit = 50L, ...)

Arguments

uploadDir
Path to final upload directory. This path is set when running "bcbio_nextgen -w template".

level
Import counts as "genes" (default) or "transcripts".

caller
Expression caller. Supports "salmon" (default), "kallisto", or "sailfish".

organism
Organism name. Use the full Latin name (e.g. "Homo sapiens"), since this will be input downstream to AnnotationHub and ensembldb, unless gffFile is set. If left NULL (not recommended), the function call will skip loading gene/transcript-level annotations into rowRanges(). This can be useful for poorly annotation genomes or experiments involving multiple genomes.

sampleMetadataFile
Optional. Custom metadata file containing sample information. Otherwise defaults to sample metadata saved in the YAML file. Remote URLs are supported. Typically this can be left unset.

samples
Optional. Specify a subset of samples to load. The names must match the description specified in the bcbio YAML metadata. If a sampleMetadataFile is provided, that will take priority for sample selection. Typically this can be left unset.

censorSamples
Optional. Samples to exclude from the analysis.

interestingGroups
Character vector denoting groups of interest that define the samples. If left unset, defaults to sampleName.

ensemblRelease
Optional. Ensembl release version. If unset, defaults to current release, and does not typically need to be user-defined. Passed to AnnotationHub for EnsDb annotation matching, unless gffFile is set.

genomeBuild
Optional. Ensembl genome build name (e.g. "GRCh38"). This will be passed to AnnotationHub for EnsDb annotation matching, unless gffFile is set.

transgeneNames
character vector indicating which assay() rows denote transgenes (e.g. EGFP, TDTOMATO).

spikeNames
character vector indicating which assay() rows denote spike-in sequences (e.g. ERCCs).

gffFile
Advanced use; not recommended. By default, we recommend leaving this NULL for genomes that are supported on Ensembl. In this case, the row annotations (rowRanges()) will be obtained automatically from Ensembl by passing the organism, genomeBuild, and ensemblRelease arguments to AnnotationHub and ensembldb. For a genome that is not supported on Ensembl and/or AnnotationHub, a GFF/GTF (General Feature Format) file is required. Generally, we recommend using a GTF (GFFv2) file here over a GFF3 file if possible, although all GFF formats are supported. The function will internally generate a TxDb containing transcript-to-gene mappings and construct a GRanges object containing the genomic ranges (rowRanges()).

transformationLimit
Maximum sample limit allowed to calculate DESeq2::rlog() and DESeq2::varianceStabilizingTransformation(). For large datasets, DESeq2 is slow to apply variance stabilization. In this case, we recommend loading up the dataset in a high-performance computing environment. Use Inf to always apply and -Inf to always skip.

...
Additional arguments, slotted into the metadata() accessor.

Value

bcbioRNASeq.

Details

Simply point to the final upload directory generated by bcbio, and this constructor function will take care of the rest. It automatically imports RNA-seq counts, metadata, and the program versions used.

This class contains raw read counts and length-scaled transcripts per million (TPM) generated by tximport::tximport(). Counts can be loaded at gene or transcript level.

Note

bcbioRNASeq extended SummarizedExperiment prior to v0.2.0, where we migrated to RangedSummarizedExperiment.

Metadata

The metadata() accessor contains:

Sample quality control metrics.

Ensembl annotations.

Server run paths.

R session information (e.g. utils::sessionInfo()).

DESeq2

DESeq2 is run automatically when bcbioRNASeq() is called, and variance stabilized counts are slotted into assays(). If the number of samples is bigger than the transformationLimit argument, rlog and vst counts will not be slotted into assays(). In this case, we recommend visualization using tmm() counts, which are automatically calculated using edgeR.

Remote data

When working in RStudio, we recommend connecting to the bcbio run directory as a remote connection over sshfs.

See also

Examples

uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")

# Gene level
x <- bcbioRNASeq(
    uploadDir = uploadDir,
    level = "genes",
    caller = "salmon",
    organism = "Mus musculus",
    ensemblRelease = 87L
)

#> 2017-05-23_rnaseq

#> 4 samples detected:
#> group1_1, group1_2, group2_1, group2_2

#> Reading project-summary.yaml

#> Reading project-summary.yaml

#> Reading sample metrics

#> Reading project-summary.yaml

#> Parsed with column specification:
#> cols(
#>   txID = col_character(),
#>   geneID = col_character()
#> )

#> Reading salmon counts using tximport

#> 1 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 2 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 3 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 4 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 

#> summarizing abundance

#> summarizing counts

#> summarizing length

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> Ensembl 87 EnsDb for Mus Musculus

#> id: "AH53222"
#> organism: "Mus musculus"
#> build: "GRCm38"
#> release: "87"
#> format: "genes"

#> Standardizing the metadata columns

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Defining broadClass using: geneName, geneBiotype, seqnames

#> Generating DESeqDataSet with DESeq2 1.20.0

#> using just counts from tximport

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

#> Applying rlog transformation

#> Applying variance stabilizing transformation

show(x)

#> bcbioRNASeq 0.2.6
#> Samples: 4
#> Genes: 502
#> Organism: Mus musculus
#> AnnotationHub: AH53222
#> Ensembl Release: 87
#> Genome Build: GRCm38
#> Upload Dir: /Users/mike/GitHub/hbc/bcbioRNASeq/inst/extdata/bcbio
#> Upload Date: 2017-05-23
#> R Load Date: 2018-07-09

is(x, "RangedSummarizedExperiment")

#> [1] TRUE

validObject(x)

#> [1] TRUE

# Transcript level
x <- bcbioRNASeq(
    uploadDir = uploadDir,
    level = "transcripts",
    caller = "salmon",
    organism = "Mus musculus",
    ensemblRelease = 87L
)

#> 2017-05-23_rnaseq

#> 4 samples detected:
#> group1_1, group1_2, group2_1, group2_2

#> Reading project-summary.yaml

#> Reading project-summary.yaml

#> Reading sample metrics

#> Reading project-summary.yaml

#> Parsed with column specification:
#> cols(
#>   txID = col_character(),
#>   geneID = col_character()
#> )

#> Reading salmon counts using tximport

#> 1 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 2 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 3 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 4 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> Ensembl 87 EnsDb for Mus Musculus

#> id: "AH53222"
#> organism: "Mus musculus"
#> build: "GRCm38"
#> release: "87"
#> format: "transcripts"

#> Standardizing the metadata columns

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Setting txBiotype as factor

#> Setting txSupportLevel as factor

#> Defining broadClass using: geneName, txBiotype, seqnames

show(x)

#> bcbioRNASeq 0.2.6
#> Samples: 4
#> Transcripts: 996
#> Organism: Mus musculus
#> AnnotationHub: AH53222
#> Ensembl Release: 87
#> Genome Build: GRCm38
#> Upload Dir: /Users/mike/GitHub/hbc/bcbioRNASeq/inst/extdata/bcbio
#> Upload Date: 2017-05-23
#> R Load Date: 2018-07-09

validObject(x)

#> [1] TRUE