bcbioRNASeq object and constructor

bcbioRNASeq is an S4 class that extends RangedSummarizedExperiment, and is designed to store a bcbio RNA-seq analysis.

bcbioRNASeq(uploadDir, level = c("genes", "transcripts"),
  caller = c("salmon", "kallisto", "sailfish", "star", "hisat2"),
  interestingGroups = "sampleName", samples = NULL,
  censorSamples = NULL, sampleMetadataFile = NULL, organism = NULL,
  genomeBuild = NULL, ensemblRelease = NULL, gffFile = NULL,
  transgeneNames = NULL, spikeNames = NULL, vst = TRUE,
  rlog = FALSE, ...)

Arguments

uploadDir
string. Path to final upload directory. This path is set when running "bcbio_nextgen -w template".

level
string. Import counts at gene level ("genes"; default) or transcript level ("transcripts"; advanced use).

caller
string. Expression caller:

"salmon" (default): Salmon alignment-free, quasi-mapped counts.

"kallisto". Kallisto alignment-free, pseudo-aligned counts.

"sailfish". Sailfish alignment-free, lightweight counts.

"star": STAR (Spliced Transcripts Alignment to a Reference) aligned counts.

"hisat2": HISAT2 (Hierarchical Indexing for Spliced Alignment of Transcripts) graph-based aligned counts.

interestingGroups
character. Groups of interest that define the samples. If left unset, defaults to sampleName.

samples
character or NULL. Specify a subset of samples to load. The names must match the description specified in the bcbio YAML metadata. If a sampleMetadataFile is provided, that will take priority for sample selection. Typically this can be left unset.

censorSamples
character or NULL. Samples to exclude from the analysis.

sampleMetadataFile
string or NULL. Custom metadata file containing sample information. Otherwise defaults to sample metadata saved in the YAML file. Remote URLs are supported. Typically this can be left unset.

organism
string or NULL. Organism name. Use the full Latin name (e.g. "Homo sapiens"), since this will be input downstream to AnnotationHub and ensembldb, unless gffFile is set. If left NULL (not recommended), the function call will skip loading gene/transcript-level annotations into rowRanges(). This can be useful for poorly annotation genomes or experiments involving multiple genomes.

genomeBuild
string or NULL. Ensembl genome build name (e.g. "GRCh38"). This will be passed to AnnotationHub for EnsDb annotation matching, unless gffFile is set.

ensemblRelease
scalar integer or NULL. Ensembl release version. If unset, defaults to current release, and does not typically need to be user-defined. Passed to AnnotationHub for EnsDb annotation matching, unless gffFile is set.

gffFile
string or NULL. By default, we recommend leaving this NULL for genomes that are supported on Ensembl. In this case, the row annotations (rowRanges()) will be obtained automatically from Ensembl by passing the organism, genomeBuild, and ensemblRelease arguments to AnnotationHub and ensembldb. For a genome that is not supported on Ensembl and/or AnnotationHub, a GFF/GTF (General Feature Format) file is required. Generally, we recommend using a GTF (GFFv2) file here over a GFF3 file if possible, although all GFF formats are supported. The function will internally generate a TxDb containing transcript-to-gene mappings and construct a GRanges object containing the genomic ranges (rowRanges()).

transgeneNames
character. Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).

spikeNames
character. Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).

vst
boolean. Calculate variance-stabilizing transformation using DESeq2::varianceStabilizingTransformation(). Recommended by default for visualization.

rlog
boolean. Calcualte regularized log transformation using DESeq2::rlog(). This calculation is slow for large datasets and now discouraged by default for visualization.

...
Additional arguments, slotted into the metadata() accessor.

Value

bcbioRNASeq.

Details

Simply point to the final upload directory generated by bcbio, and this constructor function will take care of the rest. It automatically imports RNA-seq counts, metadata, and the program versions used.

This class contains raw read counts and length-scaled transcripts per million (TPM) generated by tximport::tximport(). Counts can be loaded at gene or transcript level.

Note

bcbioRNASeq extended SummarizedExperiment prior to v0.2.0, where we migrated to RangedSummarizedExperiment.

Genome build

Ensure that the organism and genome build used with bcio match correctly here in the function call. In particular, for the legacy Homo sapiens GRCh37/hg19 genome build, ensure that genomeBuild = "GRCh37". Otherwise, the genomic ranges set in rowRanges() will mismatch. It is recommended for current projects that GRCh38/hg38 is used in place of GRCh37/hg19 if possible.

Metadata

The metadata() slot contains:

Sample quality control metrics.

Ensembl annotations.

Server run paths.

R session information (e.g. utils::sessionInfo()).

DESeq2

DESeq2 is run automatically when bcbioRNASeq() is called, and variance stabilized counts are slotted into assays(). If the number of samples is bigger than the transformationLimit argument, rlog and vst counts will not be slotted into assays(). In this case, we recommend visualization using TMM counts, which are automatically calculated using edgeR.

Remote data

When working in RStudio, we recommend connecting to the bcbio run directory as a remote connection over sshfs.

See also

Examples

upload_dir <- system.file("extdata/bcbio", package = "bcbioRNASeq")

# Gene level
x <- bcbioRNASeq(
    uploadDir = upload_dir,
    level = "genes",
    caller = "salmon",
    organism = "Mus musculus",
    ensemblRelease = 87L
)

#> 2017-05-23_rnaseq

#> 4 samples detected:
#> [1] "group1_1" "group1_2" "group2_1" "group2_2"

#> Importing project-summary.yaml using yaml::yaml.load_file().

#> Data versions are missing.

#> Importing bcbio-nextgen.log using readr::read_lines().

#> Importing bcbio-nextgen-commands.log using readr::read_lines().

#> Importing project-summary.yaml using yaml::yaml.load_file().

#> Getting sample metadata from YAML.

#> Importing project-summary.yaml using yaml::yaml.load_file().

#> Getting sample metrics from YAML.

#> Reading salmon counts using tximport 1.12.0

#> 1 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 2 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 3 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 4 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 

#> summarizing abundance

#> summarizing counts

#> summarizing length

#> Using `makeGRangesFromEnsembl()` for annotations

#> Making GRanges from Ensembl.

#> Matching EnsDb from AnnotationHub 2.16.0 (2019-05-01).

#> AH53222: Ensembl 87 EnsDb for Mus Musculus
#> Run this code to download EnsDb manually:
#>   > library(AnnotationHub)
#>   > ah <- AnnotationHub()
#>   > edb <- ah[["AH53222"]]

#> Making GRanges from EnsDb object.

#>   - Organism: Mus musculus
#>   - Genome Build: GRCm38
#>   - Ensembl Release: 87
#>   - Level: genes

#> Defining broadClass using: geneBiotype, geneName, seqnames

#> Arranging by geneID.

#> 50143 genes detected.

#> Generating DESeqDataSet with DESeq2 1.24.0

#> using just counts from tximport

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

#> Applying variance-stabilizing transformation

show(x)

#> bcbioRNASeq 0.2.10
#> class: RangedSummarizedExperiment 
#> dim: 502 4 
#> metadata(28): version level ... wd sessionInfo
#> assays(5): counts tpm length normalized vst
#> rownames(502): ENSMUSG00000002459 ENSMUSG00000004768 ...
#>   ENSMUSG00000104523 ENSMUSG00000105982
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(4): group1_1 group1_2 group2_1 group2_2
#> colData names(21): sampleName description ... totalReads x5x3Bias
#> Upload Dir: "/Users/mike/git/hbc/bcbioRNASeq/inst/extdata/bcbio"
#> Upload Date: 2017-05-23
#> R Load Date: 2019-05-13
#> Level: "genes"
#> Caller: "salmon"
#> Organism: "Mus musculus"
#> Interesting Groups: "sampleName"
#> AnnotationHub: character(0)
#> Ensembl Release: "87"
#> Genome Build: "GRCm38"

is(x, "RangedSummarizedExperiment")

#> [1] TRUE

validObject(x)

#> [1] TRUE

# Transcript level
x <- bcbioRNASeq(
    uploadDir = upload_dir,
    level = "transcripts",
    caller = "salmon",
    organism = "Mus musculus",
    ensemblRelease = 87L
)

#> 2017-05-23_rnaseq

#> 4 samples detected:
#> [1] "group1_1" "group1_2" "group2_1" "group2_2"

#> Importing project-summary.yaml using yaml::yaml.load_file().

#> Data versions are missing.

#> Importing bcbio-nextgen.log using readr::read_lines().

#> Importing bcbio-nextgen-commands.log using readr::read_lines().

#> Importing project-summary.yaml using yaml::yaml.load_file().

#> Getting sample metadata from YAML.

#> Importing project-summary.yaml using yaml::yaml.load_file().

#> Getting sample metrics from YAML.

#> Reading salmon counts using tximport 1.12.0

#> 1 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 2 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 3 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 4 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 

#> Using `makeGRangesFromEnsembl()` for annotations

#> Making GRanges from Ensembl.

#> Matching EnsDb from AnnotationHub 2.16.0 (2019-05-01).

#> AH53222: Ensembl 87 EnsDb for Mus Musculus
#> Run this code to download EnsDb manually:
#>   > library(AnnotationHub)
#>   > ah <- AnnotationHub()
#>   > edb <- ah[["AH53222"]]

#> Making GRanges from EnsDb object.

#>   - Organism: Mus musculus
#>   - Genome Build: GRCm38
#>   - Ensembl Release: 87
#>   - Level: transcripts

#> Merging gene-level annotations: gene_name, gene_biotype, seq_coord_system, description, symbol, entrezid

#> Defining broadClass using: geneName, seqnames, transcriptBiotype

#> Arranging by transcriptID.

#> 124168 transcripts detected.

show(x)

#> bcbioRNASeq 0.2.10
#> class: RangedSummarizedExperiment 
#> dim: 996 4 
#> metadata(28): version level ... wd sessionInfo
#> assays(3): counts tpm length
#> rownames(996): ENSMUST00000002533 ENSMUST00000006037 ...
#>   ENSMUST00000208793 ENSMUST00000209020
#> rowData names(13): broadClass description ... transcriptName
#>   transcriptSupportLevel
#> colnames(4): group1_1 group1_2 group2_1 group2_2
#> colData names(21): sampleName description ... totalReads x5x3Bias
#> Upload Dir: "/Users/mike/git/hbc/bcbioRNASeq/inst/extdata/bcbio"
#> Upload Date: 2017-05-23
#> R Load Date: 2019-05-13
#> Level: "transcripts"
#> Caller: "salmon"
#> Organism: "Mus musculus"
#> Interesting Groups: "sampleName"
#> AnnotationHub: character(0)
#> Ensembl Release: "87"
#> Genome Build: "GRCm38"

validObject(x)

#> [1] TRUE