bcbio RNA-Seq data set

Simply point to the final upload directory generated by bcbio, and this generator function will take care of the rest.

bcbioRNASeq(uploadDir, level = c("genes", "transcripts"),
  caller = c("salmon", "kallisto", "sailfish", "star", "hisat2"),
  samples = NULL, censorSamples = NULL, sampleMetadataFile = NULL,
  organism = NULL, genomeBuild = NULL, ensemblRelease = NULL,
  gffFile = NULL, transgeneNames = NULL, spikeNames = NULL,
  countsFromAbundance = "lengthScaledTPM", vst = TRUE, rlog = FALSE,
  interestingGroups = "sampleName", ...)

Arguments

uploadDir
character(1). Path to final upload directory. This path is set when running "bcbio_nextgen -w template".

level
character(1). Import counts at gene level ("genes"; default) or transcript level ("transcripts"; advanced use). Only tximport-compatible callers (e.g. salmon, kallisto, sailfish) can be loaded at transcript level. Aligned counts from featureCounts-compatible callers (e.g. STAR, HISAT2) can only be loaded at gene level.

caller
character(1). Expression caller:

- "`salmon`" (*default*): [Salmon][] alignment-free, quasi-mapped counts.
- "`kallisto`": [Kallisto][] alignment-free, pseudo-aligned counts.
- "`sailfish`": [Sailfish][] alignment-free, lightweight counts.
- "`star`": [STAR][] (Spliced Transcripts Alignment to a Reference)
  aligned counts.
- "`hisat2`": [HISAT2][] (Hierarchical Indexing for Spliced Alignment of
  Transcripts) graph-based aligned counts.

samples
character. Specify a subset of samples to include.

censorSamples
character. Specify a subset of samples to censor.

sampleMetadataFile
character(1). Sample metadata file path. CSV or TSV is preferred, but Excel worksheets are also supported. Check the documentation for conventions and required columns.

organism
character(1). Full Latin organism name (e.g. "Homo sapiens").

genomeBuild
character(1). Ensembl genome build assembly name (e.g. "GRCh38"). If set NULL, defaults to the most recent build available. Note: don't pass in UCSC build IDs (e.g. "hg38").

ensemblRelease
integer(1). Ensembl release version (e.g. 90). We recommend setting this value if possible, for improved reproducibility. When left unset, the latest release available via AnnotationHub/ensembldb is used. Note that the latest version available can vary, depending on the versions of AnnotationHub and ensembldb in use.

gffFile
character(1). GFF/GTF (General Feature Format) file. Generally, we recommend using a GTF (GFFv2) instead of a GFFv3 file if possible.

transgeneNames
character. Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).

spikeNames
character. Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).

countsFromAbundance
character(1). Whether to generate estimated counts using abundance estimates (recommended by default). lengthScaledTPM is a suitable default, and counts are scaled using the average transcript length over samples and then the library size. Refer to tximport::tximport() for more information on this parameter, but it should only ever be changed when loading some datasets at transcript level (e.g. for DTU analsyis).

vst
logical(1). Calculate variance-stabilizing transformation using DESeq2::varianceStabilizingTransformation(). Recommended by default for visualization.

rlog
logical(1). Calcualte regularized log transformation using DESeq2::rlog(). This calculation is slow for large datasets and now discouraged by default for visualization.

interestingGroups
character. Groups of interest that define the samples. If left unset, defaults to sampleName.

...
Additional arguments.

Value

bcbioRNASeq.

Details

The bcbioRNASeq generator function automatically imports RNA-seq counts, metadata, and the program versions used from a bcbio RNA-seq run.

Sample metadata

When loading a bcbio RNA-seq run, the sample metadata will be imported automatically from the project-summary.yaml file in the final upload directory. If you notice any typos in your metadata after completing the run, these can be corrected by editing the YAML file.

Alternatively, you can pass in a sample metadata file into the bcbioRNASeq function call using the sampleMetadataFile argument. This requires either a CSV or Excel spreadsheet.

The samples in the bcbio run must map to the description column. The values provided in description must be unique. These values will be sanitized into syntactically valid names (see basejump::makeNames for more information), and assigned as the column names of the bcbioRNASeq object. The original values are stored as the sampleName column in colData, and are used for all plotting functions. Do not attempt to set a sampleID column, as this is used internally by the package.

Here is a minimal example of a properly formatted sample metadata file:

description

genotype

sample1	wildtype

sample2

knockout

sample3	wildtype

sample4

knockout

Valid names

R is strict about values that are considered valid for use in names and dimnames (i.e. rownames and colnames). Non-alphanumeric characters, spaces, and dashes are not valid. Use either underscores or periods in place of dashes when working in R. Also note that names should not begin with a number, and will be prefixed with an X when sanitized. Consult the documentation in the make.names function for more information. We strongly recommend adhering to these conventions when labeling samples, to help avoid unexpected downstream behavior in R due to dimnames mismatches.

Genome annotations

bcbioRNASeq provides support for automatic import of genome annotations, which internally get processed into genomic ranges (GRanges) and are slotted into the rowRanges of the S4 object. Currently, we offer support for (1) Ensembl genome annotations from AnnotationHub via ensembldb (recommended); or (2) direct import from a GTF/GFF file using rtracklayer.

ensembldb requires the organism and ensemblRelease arguments to be defined. When both of these are set, bcbioRNASeq will attempt to download and use the pre-built Ensembl genome annotations from AnnotationHub. This method is preferred over direct loading of a GTF/GFF file because the AnnotationHub annotations contain additional rich metadata not defined in a GFF file, specifically description and entrezID values.

Alternatively, if you are working with a non-standard or poorly annotated genome that isn't available on AnnotationHub, we provide fall back support for loading the genome annotations directly from the GTF file used by the bcbio RNA-seq pipeline. This should be fully automatic for an R session active on the same server used to run bcbio.

Example bcbio GTF path: genomes/Hsapiens/hg38/rnaseq/ref-transcripts.gtf.

In the event that you are working from a remote environment that doesn't have file system access to the bcbio genomes directory, we provide additional fall back support for importing genome annotations from a GTF/GFF directly with the gffFile argument.

Internally, genome annotations are imported via the basejump package, specifically with either of these functions:

basejump::makeGRangesFromEnsembl().

basejump::makeGRangesFromGFF().

Genome build

Ensure that the organism and genome build used with bcio match correctly here in the function call. In particular, for the legacy Homo sapiens GRCh37/hg19 genome build, ensure that genomeBuild = "GRCh37". Otherwise, the genomic ranges set in rowRanges will mismatch. It is recommended for current projects that GRCh38/hg38 is used in place of GRCh37/hg19 if possible.

DESeq2

DESeq2 is run automatically when bcbioRNASeq() is called. Internally, this automatically slots normalized counts into assays(), and optionally generates variance-stabilized counts (e.g. vst, rlog), depending on the call. When loading a dataset with a large number of samples (i.e. > 50), we recommend disabling the rlog transformation, since it can take a long time to compute.

Remote connections

When working on a local machine, it is possible to load bcbio run data over a remote connection using sshfs. When loading a large number of samples, it is preferable to call bcbioRNASeq() directly in R on the remote server, if possible.

Automatic metadata

The metadata slot always contains:

Object version.

bcbio data provenance information.

File paths and timestamps.

R session information.

See also

Examples

uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")

## Gene level.
object <- bcbioRNASeq(
    uploadDir = uploadDir,
    level = "genes",
    caller = "salmon",
    organism = "Mus musculus",
    ensemblRelease = 87L
)

#> Dated project directory: 2018-03-18_GSE65267-merged

#> 6 sample(s) detected:
#> control_rep1, control_rep2, control_rep3, fa_day7_rep1, fa_day7_rep2, fa_day7...

#> Importing project-summary.yaml using yaml::yaml.load_file().

#> Data versions are missing.

#> Importing bcbio-nextgen.log using readr::read_lines().

#> bcbio-nextgen.log file is empty.

#> Importing bcbio-nextgen-commands.log using readr::read_lines().

#> bcbio-nextgen-commands.log file is empty.

#> Getting sample metadata from YAML.

#> Getting sample metrics from YAML.

#> Reading salmon counts using tximport 1.10.1.

#> Reading from quant.sf files.

#> Scaling using lengthScaledTPM.

#> 1 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 2 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 3 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 4 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 5 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 6 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 

#> summarizing abundance

#> summarizing counts

#> summarizing length

#> Using makeGRangesFromEnsembl() for annotations.

#> Making GRanges from Ensembl.

#> Matching EnsDb from AnnotationHub 2.14.2 (2018-10-24).

#> AH53222: Ensembl 87 EnsDb for Mus Musculus

#> Making GRanges from EnsDb object.

#>   - Organism: Mus musculus
#>   - Genome Build: GRCm38
#>   - Ensembl Release: 87
#>   - Level: genes

#> Using `geneName` column instead of `symbol`.

#> Defining broadClass using: geneName, geneBiotype, seqnames

#> Arranging by geneID.

#> Generating DESeqDataSet with DESeq2 1.22.2.

#> Calculating normalized counts.

#> Calculating transformations.

#> using pre-existing size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

#> Applying variance-stabilizing transformation.

#> Calculating FPKM.

print(object)

#> bcbioRNASeq 0.3.7
#> uploadDir: /Users/mike/git/bcbioRNASeq/inst/extdata/bcbio
#> dates(2): [bcbio] 2018-03-18; [R] 2019-01-13
#> level: genes
#> caller: salmon
#> organism: Mus musculus
#> interestingGroups: sampleName
#> class: RangedSummarizedExperiment 
#> dim: 100 6 
#> metadata(26): version level ... wd sessionInfo
#> assays(6): counts tpm ... vst fpkm
#> rownames(100): ENSMUSG00000000001 ENSMUSG00000000003 ...
#>   ENSMUSG00000062661 ENSMUSG00000074340
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(6): control_rep1 control_rep2 ... fa_day7_rep2 fa_day7_rep3
#> colData names(26): averageInsertSize averageReadLength ... tissue
#>   treatment

## Transcript level.
object <- bcbioRNASeq(
    uploadDir = uploadDir,
    level = "transcripts",
    caller = "salmon",
    organism = "Mus musculus",
    ensemblRelease = 87L
)

#> Dated project directory: 2018-03-18_GSE65267-merged

#> 6 sample(s) detected:
#> control_rep1, control_rep2, control_rep3, fa_day7_rep1, fa_day7_rep2, fa_day7...

#> Importing project-summary.yaml using yaml::yaml.load_file().

#> Data versions are missing.

#> Importing bcbio-nextgen.log using readr::read_lines().

#> bcbio-nextgen.log file is empty.

#> Importing bcbio-nextgen-commands.log using readr::read_lines().

#> bcbio-nextgen-commands.log file is empty.

#> Getting sample metadata from YAML.

#> Getting sample metrics from YAML.

#> Reading salmon counts using tximport 1.10.1.

#> Reading from quant.sf files.

#> Scaling using lengthScaledTPM.

#> 1 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 2 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 3 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 4 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 5 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 6 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_double(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 

#> Using makeGRangesFromEnsembl() for annotations.

#> Making GRanges from Ensembl.

#> Matching EnsDb from AnnotationHub 2.14.2 (2018-10-24).

#> AH53222: Ensembl 87 EnsDb for Mus Musculus

#> Making GRanges from EnsDb object.

#>   - Organism: Mus musculus
#>   - Genome Build: GRCm38
#>   - Ensembl Release: 87
#>   - Level: transcripts

#> Using `geneName` column instead of `symbol`.

#> Defining broadClass using: geneName, transcriptBiotype, seqnames

#> Arranging by transcriptID.

print(object)

#> bcbioRNASeq 0.3.7
#> uploadDir: /Users/mike/git/bcbioRNASeq/inst/extdata/bcbio
#> dates(2): [bcbio] 2018-03-18; [R] 2019-01-13
#> level: transcripts
#> caller: salmon
#> organism: Mus musculus
#> interestingGroups: sampleName
#> class: RangedSummarizedExperiment 
#> dim: 100 6 
#> metadata(26): version level ... wd sessionInfo
#> assays(3): counts tpm avgTxLength
#> rownames(100): ENSMUST00000000001 ENSMUST00000000003 ...
#>   ENSMUST00000000674 ENSMUST00000000687
#> rowData names(13): broadClass description ... transcriptName
#>   transcriptSupportLevel
#> colnames(6): control_rep1 control_rep2 ... fa_day7_rep2 fa_day7_rep3
#> colData names(26): averageInsertSize averageReadLength ... tissue
#>   treatment