bcbioRNASeq Object Class
Source: R/AllClasses.R

S4 class containing RNA-seq counts and metadata generated by bcbio.

Details

bcbioRNASeq is a subclass of SummarizedExperiment designed to store an RNA-seq analysis. This class contains raw read counts and transcripts per million (TPM) generated by tximport().

The metadata accessor contains:

Sample summary statistics (a.k.a. metrics).

Ensembl annotations.

Server run paths.

R local environment information, including sessionInfo().

Slots

bcbio

SimpleList containing additional bcbio run data with dimensions that don't match the count matrix. This is currently used to store STAR featureCounts aligned counts.

See also

Examples

uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")

# We're suppressing warnings about missing log files in our minimal
# dataset, which end up in the Travis CI logs. This isn't recommended
# for a typical bcbio run.
bcb <- loadRNASeq(uploadDir)

#> 2017-05-23_rnaseq

#> 4 samples detected

#> Reading project-summary.yaml

#> Detecting organism from genome build

#> Genome: Mus musculus (mm10)

#> Loading Ensembl annotations from AnnotationHub
#> 2017-10-27

#> EnsDB AH57770: Mus musculus Ensembl 90

#> Obtaining transcript-to-gene mappings

#> Parsed with column specification:
#> cols(
#>   enstxp = col_character(),
#>   ensgene = col_character()
#> )

#> Reading sample metrics

#> Reading bcbio run information

#> Parsed with column specification:
#> cols(
#>   genome = col_character(),
#>   resource = col_character(),
#>   version = col_datetime(format = "")
#> )

#> bcbio-nextgen.log.txt

#> bcbio-nextgen-commands.log.txt

#> Reading salmon counts using tximport

#> 1 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 2 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 3 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 4 

#> Parsed with column specification:
#> cols(
#>   Name = col_character(),
#>   Length = col_integer(),
#>   EffectiveLength = col_double(),
#>   TPM = col_double(),
#>   NumReads = col_double()
#> )

#> 

#> summarizing abundance

#> summarizing counts

#> summarizing length

#> Performing trimmed mean of M-values (TMM) normalization

#> Generating internal DESeqDataSet

#> using just counts from tximport

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

#> Performing rlog transformation

#> Performing variance stabilizing transformation

#> Reading STAR featureCounts aligned counts

#> Parsed with column specification:
#> cols(
#>   id = col_character(),
#>   group1_1 = col_integer(),
#>   group1_2 = col_integer(),
#>   group2_1 = col_integer(),
#>   group2_2 = col_integer()
#> )

#> Warning: Unannotated genes detected in assay (0.594%)

print(bcb)

#> class: bcbioRNASeq 
#> dim: 505 4 
#> metadata(28): version uploadDir ... devtoolsSessionInfo
#>   unannotatedGenes
#> assays(6): raw normalized ... rlog vst
#> rownames(505): ENSMUSG00000002459 ENSMUSG00000004768 ...
#>   ENSMUSG00000105982 ENSMUSG00000109048
#> rowData names(11): ensgene symbol ... seqCoordSystem entrez
#> colnames(4): group1_1 group1_2 group2_1 group2_2
#> colData names(4): sampleID sampleName description group

# Check validity
validObject(bcb)

#> [1] TRUE