Counts

Count matrix.

# S4 method for bcbioRNASeq
counts(object, normalized = FALSE)

Arguments

object
Object.

normalized
character(1) or logical(1). Which normalization method to apply:

FALSE: Raw counts. When using a tximport-compatible caller, these are length scaled by default (see countsFromAbundance argument). When using a featureCounts-compatible caller, these are integer.

tximport caller-specific normalizations:

"tpm": Transcripts per million.

Additional gene-level-specific normalizations:

TRUE: Size factor-adjusted counts.
See DESeq2::sizeFactors for more information.

"vst": Variance-stabilizing transformation.
Requires vst = TRUE to be set during bcbioRNASeq call.
See DESeq2::varianceStabilizingTransformation for more information.

"rlog": Regularized log transformation (log2).
Requires rlog = TRUE to be set during bcbioRNASeq call.
See DESeq2::rlog for more information.

"tmm": Trimmed mean of M-values.
See edgeR::calcNormFactors for more information.

"rle": Relative log expression transformation.

"fpkm": Fragments per kilobase per million mapped fragments.
Requires annotations in rowRanges with defined width; otherwise, will be skipped during the bcbioRNASeq load call.
See BiocGenerics::width and DESeq2::fpkm for more information.

Value

matrix.

Details

By default, return the raw counts. Normalized counts in a variety of formats can be accessed using the normalized argument.

References

TMM: Robinson and Oshlack (2010).

RLE: Anders and Huber (2010).

See also

Examples

data(bcb)
summary(counts(bcb))

#>   control_rep1       control_rep2       control_rep3       fa_day7_rep1     
#>  Min.   :       0   Min.   :       0   Min.   :       0   Min.   :       1  
#>  1st Qu.:     419   1st Qu.:     904   1st Qu.:     317   1st Qu.:     579  
#>  Median :    1389   Median :    3397   Median :    1315   Median :    1725  
#>  Mean   :   43568   Mean   :   86384   Mean   :   41167   Mean   :   39351  
#>  3rd Qu.:    4692   3rd Qu.:   11303   3rd Qu.:    4125   3rd Qu.:    5373  
#>  Max.   :11767239   Max.   :22627793   Max.   :11608372   Max.   :10623916  
#>   fa_day7_rep2      fa_day7_rep3     
#>  Min.   :      0   Min.   :       0  
#>  1st Qu.:    501   1st Qu.:     769  
#>  Median :   1557   Median :    1904  
#>  Mean   :  34610   Mean   :   42438  
#>  3rd Qu.:   4840   3rd Qu.:    5304  
#>  Max.   :9676254   Max.   :12145225