Load bcbio RNA-Seq Data

Simply point to the final upload directory output by bcbio, and this function will take care of the rest. It automatically imports RNA-seq counts, metadata, and program versions used.

loadRNASeq(uploadDir, interestingGroups = "sampleName",
  sampleMetadataFile = NULL, maxSamples = 50, ensemblVersion = "current",
  ...)

Arguments

uploadDir
Path to final upload directory. This path is set when running bcbio_nextgen -w template.

interestingGroups
Character vector of interesting groups. First entry is used for plot colors during quality control (QC) analysis. Entire vector is used for PCA and heatmap QC functions.

sampleMetadataFile
Optional. Custom metadata file containing sample information. Otherwise defaults to sample metadata saved in the YAML file.

maxSamples
Maximum number of samples to calculate DESeq2::rlog() and DESeq2::varianceStabilizingTransformation() matrix. See Details.

ensemblVersion
Ensembl release version. Defaults to current, and does not typically need to be user-defined. This parameter can be useful for matching Ensembl annotations against an outdated bcbio annotation build.

...
Additional arguments, slotted into the metadata() accessor.

Value

bcbioRNASeq.

Details

When number of samples is bigger than maxSamples, rlog and vst slot in SummarizedExperiment::SummarizedExperiment will be the output of edgeR normalization method.

Note

When working in RStudio, we recommend connecting to the bcbio-nextgen run directory as a remote connection over sshfs.

Examples

uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")
bcb <- loadRNASeq(uploadDir, interestingGroups = "group")

#> 2017-05-23_rnaseq

#> 4 samples detected

#> Reading project-summary.yaml

#> Genome: Mus musculus (mm10)

#> Loading Ensembl annotations from AnnotationHub

#> /Users/mike/.AnnotationHub

#> 2017-10-18

#> EnsDB AH57770: Mus musculus Ensembl 90

#> Parsed with column specification:
#> cols(
#>   enstxp = col_character(),
#>   ensgene = col_character()
#> )

#> Reading bcbio run information

#> Error: `file` must be a string, raw vector or a connection.