Differential Expression Results Tables

Differential Expression Results Tables

resultsTables(object, ...)

# S4 method for DESeqResults
resultsTables(object, lfc = 0L, annotable = NULL,
  summary = TRUE, headerLevel = 2L, write = FALSE, dir = getwd(),
  dropboxDir = NULL, rdsToken = NA, quiet = FALSE)

Arguments

object
Object.

...
Additional arguments.

lfc
Log fold change ratio (base 2) cutoff. Does not apply to statistical hypothesis testing, only gene filtering in the results tables. See results() for additional information about using lfcThreshold and altHypothesis to set an alternative hypothesis based on expected fold changes.

annotable
Join Ensembl gene annotations to the results. Apply gene identifier to symbol mappings. A previously saved annotable data.frame is recommended. Alternatively if set NULL, then gene annotations will not be added to the results.

summary
Show summary statistics.

headerLevel
R Markdown header level. Applies only when summary = TRUE.

write
Write CSV files to disk.

dir
Local directory path where to write the results tables.

dropboxDir
Dropbox directory path where to archive the results tables for permanent storage (e.g. Stem Cell Commons). When this option is enabled, unique links per file are generated internally with the rdrop2 package.

rdsToken
RDS file token to use for Dropbox authentication.

quiet
If TRUE, suppress any status messages and/or progress bars.

Value

Results list.

Examples

load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))
load(system.file("extdata/res.rda", package = "bcbioRNASeq"))

# Can use a stashed annotable for improved speed
annotable <- annotable(bcb)

resTbl <- resultsTables(
    res,
    lfc = 0.25,
    annotable = annotable,
    summary = TRUE,
    headerLevel = 2L,
    write = TRUE,
    dir = "resultsTables",
    dropboxDir = file.path("bcbioRNASeq_examples", "resultsTables"),
    rdsToken = system.file("extdata/token.rds", package = "bcbioRNASeq")
)

#> 
#> 
#> ## Summary statistics
#> 
#> 
#> - 505 genes in counts matrix
#> - Base mean > 0: 304 genes (non-zero)
#> - Base mean > 1: 213 genes
#> - Alpha cutoff: 0.1
#> - LFC cutoff: 0.25 (applied in tables only)
#> - DEG pass alpha: 4 genes
#> - DEG LFC up: 2 genes
#> - DEG LFC down: 2 genes
#> 

#> Dropbox: Michael Steinbaugh <mjsteinbaugh@protonmail.com>

#> /bcbioRNASeq_examples/resultsTables/group_ko_vs_ctrl_all.csv.gz was successfully deleted

#> File /Users/mike/GitHub/hbc/bcbioRNASeq/docs/reference/resultsTables/group_ko_vs_ctrl_all.csv.gz uploaded as /bcbioRNASeq_examples/resultsTables/group_ko_vs_ctrl_all.csv.gz successfully at 2018-02-19T22:15:51Z

#> /bcbioRNASeq_examples/resultsTables/group_ko_vs_ctrl_deg.csv.gz was successfully deleted

#> File /Users/mike/GitHub/hbc/bcbioRNASeq/docs/reference/resultsTables/group_ko_vs_ctrl_deg.csv.gz uploaded as /bcbioRNASeq_examples/resultsTables/group_ko_vs_ctrl_deg.csv.gz successfully at 2018-02-19T22:15:54Z

#> /bcbioRNASeq_examples/resultsTables/group_ko_vs_ctrl_deg_lfc_up.csv.gz was successfully deleted

#> File /Users/mike/GitHub/hbc/bcbioRNASeq/docs/reference/resultsTables/group_ko_vs_ctrl_deg_lfc_up.csv.gz uploaded as /bcbioRNASeq_examples/resultsTables/group_ko_vs_ctrl_deg_lfc_up.csv.gz successfully at 2018-02-19T22:15:58Z

#> /bcbioRNASeq_examples/resultsTables/group_ko_vs_ctrl_deg_lfc_down.csv.gz was successfully deleted

#> File /Users/mike/GitHub/hbc/bcbioRNASeq/docs/reference/resultsTables/group_ko_vs_ctrl_deg_lfc_down.csv.gz uploaded as /bcbioRNASeq_examples/resultsTables/group_ko_vs_ctrl_deg_lfc_down.csv.gz successfully at 2018-02-19T22:16:03Z

#> 
#> 
#> ## Results tables
#> 
#> 
#> - [`group_ko_vs_ctrl_all.csv.gz`](https://www.dropbox.com/s/upev0uc9e9qj0ny/group_ko_vs_ctrl_all.csv.gz?dl=0): All genes, sorted by Ensembl identifier.
#> - [`group_ko_vs_ctrl_deg.csv.gz`](https://www.dropbox.com/s/4416e9a7422g7yx/group_ko_vs_ctrl_deg.csv.gz?dl=0): Genes that pass the alpha (FDR) cutoff.
#> - [`group_ko_vs_ctrl_deg_lfc_up.csv.gz`](https://www.dropbox.com/s/mdr0h0ofbc68nbo/group_ko_vs_ctrl_deg_lfc_up.csv.gz?dl=0): Upregulated DEG; positive log2 fold change.
#> - [`group_ko_vs_ctrl_deg_lfc_down.csv.gz`](https://www.dropbox.com/s/dj2inosd3y1aln4/group_ko_vs_ctrl_deg_lfc_down.csv.gz?dl=0): Downregulated DEG; negative log2 fold change.
#> 

names(resTbl)

#>  [1] "contrast"     "alpha"        "lfc"          "all"          "deg"         
#>  [6] "degLFC"       "degLFCUp"     "degLFCDown"   "localFiles"   "dropboxFiles"

dir("resultsTables")

#> [1] "group_ko_vs_ctrl_all.csv.gz"         
#> [2] "group_ko_vs_ctrl_deg_lfc_down.csv.gz"
#> [3] "group_ko_vs_ctrl_deg_lfc_up.csv.gz"  
#> [4] "group_ko_vs_ctrl_deg.csv.gz"         

unlink("resultsTables", recursive = TRUE)