Differential Expression Results Tables

Differential Expression Results Tables

resultsTables(bcb, res, lfc = 0L, write = TRUE, dir = file.path("results",
  "de"))

Arguments

bcb
bcbioRNADataSet.

res
DESeqResults.

lfc
Log fold change ratio (base 2) cutoff. Does not apply to statistical hypothesis testing, only gene filtering in the results tables. See results() for additional information about using lfcThreshold and altHypothesis to set an alternative hypothesis based on expected fold changes.

write
Write CSV files to disk.

dir
Directory path where to write files.

Value

Results list.

Examples

# NOT RUN {
resultsTables(bcb, res, lfc = 0.25)
# }