Differential Expression Results Tables

Differential Expression Results Tables

resultsTables(object, ...)

# S4 method for DESeqResults
resultsTables(object, lfc = 0L, rowData = NULL,
  summary = TRUE, headerLevel = 2L, write = FALSE, dir = ".",
  dropboxDir = NULL, rdsToken = NULL, ...)

Arguments

object
Object.

...
Additional arguments.

lfc
Log fold change ratio (base 2) cutoff.

rowData
Join Ensembl gene annotations to the results. Apply gene identifier to symbol mappings. A previously saved data.frame is recommended. Alternatively if set NULL, then gene annotations will not be added to the results.

summary
Show summary statistics.

headerLevel
R Markdown header level.

write
Write CSV files to disk.

dir
Local directory path where to write the results tables.

dropboxDir
Dropbox directory path where to archive the results tables for permanent storage (e.g. Stem Cell Commons). When this option is enabled, unique links per file are generated internally with the rdrop2 package.

rdsToken
RDS file token to use for Dropbox authentication.

Value

Results list.

Note

Log fold change cutoff ("lfc") does not apply to statistical hypothesis testing, only gene filtering in the results tables. See DESeq2::results() for additional information about using lfcThreshold and altHypothesis to set an alternative hypothesis based on expected fold changes.

See also

Other Differential Expression Functions: alphaSummary, contrastName, plotDEGHeatmap, plotDEGPCA, plotMA, plotVolcano

Examples

load(system.file("extdata/bcb_small.rda", package = "bcbioRNASeq"))
load(system.file("extdata/res_small.rda", package = "bcbioRNASeq"))

# DESeqResults ====
resTbl <- resultsTables(
    object = res_small,
    lfc = 0.25,
    rowData = rowData(bcb_small),
    summary = TRUE,
    headerLevel = 2L,
    write = FALSE
)

#> 
#> 
#> ## Summary statistics
#> 
#> 
#> - 500 genes in counts matrix
#> - Base mean > 0: 500 genes (non-zero)
#> - Base mean > 1: 500 genes
#> - Alpha cutoff: 0.1
#> - LFC cutoff: 0.25 (applied in tables only)
#> - DEG pass alpha: 277 genes
#> - DEG LFC up: 131 genes
#> - DEG LFC down: 146 genes
#> 

names(resTbl)

#> [1] "contrast"   "alpha"      "lfc"        "all"        "deg"       
#> [6] "degLFC"     "degLFCUp"   "degLFCDown"