Differential expression results tables

Differential expression results tables

resultsTables(results, counts, ...)

# S4 method for DESeqResults,DESeqDataSet
resultsTables(results, counts, alpha,
  lfcThreshold = 0L, summary = TRUE, write = FALSE,
  headerLevel = 2L, dir = ".", dropboxDir = NULL, rdsToken = NULL)

# S4 method for DESeqResults,missingOrNULL
resultsTables(results,
  counts = NULL, alpha, lfcThreshold = 0L)

Arguments

results
DESeqResults.

counts
DESeqDataSet. Corresponding object containing the counts that were used to generate the DESeqResults object.

...
Additional arguments.

alpha
scalar numeric. Adjusted P value ("alpha") cutoff.

lfcThreshold
scalar numeric. Log fold change ratio (base 2) cutoff threshold.

summary
boolean. Show summary statistics.

write
boolean. Write CSV files to disk.

headerLevel
scalar integer (1-7). R Markdown header level.

dir
string. Local directory path.

dropboxDir
string or NULL. Dropbox directory path where to archive the results tables for permanent storage (e.g. Stem Cell Commons). When this option is enabled, unique links per file are generated internally with the rdrop2 package. Note that local files are written to base::tempdir() and the dir argument is ignored, if this is enabled.

rdsToken
string or NULL. RDS file token to use for Dropbox authentication. If set NULL and dropboxDir is defined, then an interactive prompt will appear requesting authorization.

Value

list containing modified DESeqResults return, including additional gene-level metadata and normalized counts.

Note

Log fold change cutoff threshold ("lfcThreshold") does not apply to statistical hypothesis testing, only gene filtering in the results tables. See DESeq2::results() for additional information about using lfcThreshold and altHypothesis to set an alternative hypothesis based on expected fold changes.

See also

Examples

# DESeqResults, DESeqDataSet ====
x <- resultsTables(results = res_small, counts = dds_small)

#> 
#> 
#> ## treatment folic acid vs control
#> 
#> 
#> 
#> ### Summary statistics
#> 
#> 
#> - 500 genes in counts matrix
#> - Base mean > 0: 500 genes (non-zero)
#> - Base mean > 1: 500 genes
#> - Alpha: 0.1
#> - LFC threshold: 0
#> - DEG pass alpha: 254 genes
#> - DEG LFC up: 115 genes
#> - DEG LFC down: 139 genes
#> 

names(x)

#> [1] "deg"          "degLFC"       "degLFCUp"     "degLFCDown"   "all"         
#> [6] "contrast"     "alpha"        "lfcThreshold"

x[["deg"]] %>% str()

#> 'data.frame':	254 obs. of  44 variables:
#>  $ baseMean                                     : num  12562 69340 3560 3544 3002 ...
#>  $ log2FoldChange                               : num  8.27 5.36 4.08 3.56 3.74 ...
#>  $ lfcSE                                        : num  0.274 0.286 0.283 0.255 0.284 ...
#>  $ stat                                         : num  29.9 18.7 14.4 14 13.2 ...
#>  $ pvalue                                       : num  7.87e-197 3.65e-78 4.35e-47 3.13e-44 1.65e-39 ...
#>  $ padj                                         : num  3.89e-194 9.03e-76 7.16e-45 3.87e-42 1.63e-37 ...
#>  $ geneID                                       : chr  "ENSMUSG00000024164" "ENSMUSG00000029304" "ENSMUSG00000003617" "ENSMUSG00000022037" ...
#>  $ seqnames                                     : Factor w/ 86 levels "3","X","16","7",..: 11 12 1 14 5 13 15 6 11 18 ...
#>  $ start                                        : int  57203970 104435118 19957054 65968483 46735080 32384633 117277331 136872435 34263209 60803848 ...
#>  $ end                                          : int  57228136 104441050 20009145 65981547 46779578 32388252 117282321 136875823 34269418 60812646 ...
#>  $ width                                        : int  24167 5933 52092 13065 44499 3620 4991 3389 6210 8799 ...
#>  $ strand                                       : Factor w/ 3 levels "+","-","*": 2 1 1 1 1 2 2 2 1 1 ...
#>  $ broadClass                                   : Factor w/ 8 levels "coding","ig",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ description                                  : Factor w/ 48505 levels " 1 isoform 3  [Source:RefSeq peptide;Acc:NP_001177633]",..: 3427 42477 2783 3125 7251 9435 9628 9919 7362 2461 ...
#>  $ geneBiotype                                  : Factor w/ 44 levels "3prime_overlapping_ncRNA",..: 22 22 22 22 22 22 22 22 22 22 ...
#>  $ geneName                                     : Factor w/ 52861 levels "0610005C13Rik",..: 8340 48497 9898 9647 35979 38133 38589 39104 35851 9012 ...
#>  $ seqCoordSystem                               : Factor w/ 2 levels "chromosome","scaffold": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ baseVar                                      : num  1.88e+08 6.00e+09 1.37e+07 1.17e+07 8.87e+06 ...
#>  $ allZero                                      : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ dispGeneEst                                  : num  0.0124 0.0533 0.0398 0.0201 0.0416 ...
#>  $ dispGeneIter                                 : num  14 3 7 8 7 8 3 8 8 4 ...
#>  $ dispFit                                      : num  0.0856 0.0839 0.0909 0.0909 0.0922 ...
#>  $ dispersion                                   : num  0.0527 0.0658 0.0628 0.0502 0.0636 ...
#>  $ dispIter                                     : int  10 9 7 11 7 10 9 12 8 5 ...
#>  $ dispOutlier                                  : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ dispMAP                                      : num  0.0527 0.0658 0.0628 0.0502 0.0636 ...
#>  $ Intercept                                    : num  5.9 11.39 8.42 8.96 8.51 ...
#>  $ treatment_folic_acid_vs_control              : num  8.72 5.66 4.31 3.72 3.95 ...
#>  $ SE_Intercept                                 : num  0.22 0.214 0.214 0.19 0.215 ...
#>  $ SE_treatment_folic_acid_vs_control           : num  0.291 0.303 0.299 0.267 0.301 ...
#>  $ WaldStatistic_Intercept                      : num  26.9 53.1 39.4 47.1 39.6 ...
#>  $ WaldStatistic_treatment_folic_acid_vs_control: num  29.9 18.7 14.4 14 13.2 ...
#>  $ WaldPvalue_Intercept                         : num  6.94e-159 0.00 0.00 0.00 0.00 ...
#>  $ WaldPvalue_treatment_folic_acid_vs_control   : num  7.87e-197 3.65e-78 4.35e-47 3.13e-44 1.65e-39 ...
#>  $ betaConv                                     : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ betaIter                                     : num  2 3 3 2 3 4 3 2 3 3 ...
#>  $ deviance                                     : num  81.6 116 85 84.9 84.3 ...
#>  $ maxCooks                                     : num  0.541 0.363 0.246 0.186 0.313 ...
#>  $ control_rep1                                 : num  78.1 2313.6 301.6 512.8 334.5 ...
#>  $ control_rep2                                 : num  46.8 2998.4 344.7 536.8 452.1 ...
#>  $ control_rep3                                 : num  56.2 2733.3 379.9 445.7 301.7 ...
#>  $ fa_day7_rep1                                 : num  25465 182255 8797 7941 4727 ...
#>  $ fa_day7_rep2                                 : num  25910 121258 6227 5552 5293 ...
#>  $ fa_day7_rep3                                 : num  23817 104478 5311 6276 6902 ...