Differential expression results tables

Differential expression results tables

resultsTables(results, counts, ...)

# S4 method for DESeqResults,DESeqDataSet
resultsTables(results, counts, alpha,
  lfcThreshold = 0L, summary = TRUE, write = FALSE,
  headerLevel = 2L, dir = ".", dropboxDir = NULL, rdsToken = NULL)

# S4 method for DESeqResults,missingOrNULL
resultsTables(results,
  counts = NULL, alpha, lfcThreshold = 0L)

Arguments

results
DESeqResults.

counts
DESeqDataSet. Corresponding object containing the counts that were used to generate the DESeqResults object.

...
Additional arguments.

alpha
scalar numeric. Adjusted P value ("alpha") cutoff.

lfcThreshold
scalar numeric. Log fold change ratio (base 2) cutoff threshold.

summary
boolean. Show summary statistics.

write
boolean. Write CSV files to disk.

headerLevel
scalar integer (1-7). R Markdown header level.

dir
string. Local directory path.

dropboxDir
string or NULL. Dropbox directory path where to archive the results tables for permanent storage (e.g. Stem Cell Commons). When this option is enabled, unique links per file are generated internally with the rdrop2 package. Note that local files are written to base::tempdir() and the dir argument is ignored, if this is enabled.

rdsToken
string or NULL. RDS file token to use for Dropbox authentication. If set NULL and dropboxDir is defined, then an interactive prompt will appear requesting authorization.

Value

list containing modified DESeqResults return, including additional gene-level metadata and normalized counts.

Note

Log fold change cutoff threshold ("lfcThreshold") does not apply to statistical hypothesis testing, only gene filtering in the results tables. See DESeq2::results() for additional information about using lfcThreshold and altHypothesis to set an alternative hypothesis based on expected fold changes.

See also

Examples

# DESeqResults, DESeqDataSet ====
x <- resultsTables(results = res_small, counts = dds_small)

#> 
#> 
#> ## treatment folic acid vs control
#> 
#> 
#> 
#> ### Summary statistics
#> 
#> 
#> - 500 genes in counts matrix
#> - Base mean > 0: 500 genes (non-zero)
#> - Base mean > 1: 500 genes
#> - Alpha: 0.1
#> - LFC threshold: 0
#> - DEG pass alpha: 254 genes
#> - DEG LFC up: 115 genes
#> - DEG LFC down: 139 genes
#> 

names(x)

#> [1] "deg"          "degLFC"       "degLFCUp"     "degLFCDown"   "all"         
#> [6] "contrast"     "alpha"        "lfcThreshold"

tibble::glimpse(x[["deg"]])

#> Observations: 254
#> Variables: 44
#> $ baseMean                                      <dbl> 12562.2432, 69339.5105,…
#> $ log2FoldChange                                <dbl> 8.271501, 5.355191, 4.0…
#> $ lfcSE                                         <dbl> 0.2736464, 0.2861448, 0…
#> $ stat                                          <dbl> 29.930599, 18.716176, 1…
#> $ pvalue                                        <dbl> 7.870297e-197, 3.654651…
#> $ padj                                          <dbl> 3.887927e-194, 9.026988…
#> $ geneID                                        <chr> "ENSMUSG00000024164", "…
#> $ seqnames                                      <fct> 17, 5, 3, 14, 11, 2, 10…
#> $ start                                         <int> 57203970, 104435118, 19…
#> $ end                                           <int> 57228136, 104441050, 20…
#> $ width                                         <int> 24167, 5933, 52092, 130…
#> $ strand                                        <fct> -, +, +, +, +, -, -, -,…
#> $ broadClass                                    <fct> coding, coding, coding,…
#> $ description                                   <fct> "complement component 3…
#> $ geneBiotype                                   <fct> protein_coding, protein…
#> $ geneName                                      <fct> C3, Spp1, Cp, Clu, Havc…
#> $ seqCoordSystem                                <fct> chromosome, chromosome,…
#> $ baseVar                                       <dbl> 188041599.0, 6002054824…
#> $ allZero                                       <lgl> FALSE, FALSE, FALSE, FA…
#> $ dispGeneEst                                   <dbl> 0.012368440, 0.05329255…
#> $ dispGeneIter                                  <dbl> 14, 3, 7, 8, 7, 8, 3, 8…
#> $ dispFit                                       <dbl> 0.08561209, 0.08390861,…
#> $ dispersion                                    <dbl> 0.05265557, 0.06580512,…
#> $ dispIter                                      <int> 10, 9, 7, 11, 7, 10, 9,…
#> $ dispOutlier                                   <lgl> FALSE, FALSE, FALSE, FA…
#> $ dispMAP                                       <dbl> 0.05265557, 0.06580512,…
#> $ Intercept                                     <dbl> 5.898032, 11.389176, 8.…
#> $ treatment_folic_acid_vs_control               <dbl> 8.715299, 5.664043, 4.3…
#> $ SE_Intercept                                  <dbl> 0.2196067, 0.2142979, 0…
#> $ SE_treatment_folic_acid_vs_control            <dbl> 0.2911836, 0.3026282, 0…
#> $ WaldStatistic_Intercept                       <dbl> 26.85725, 53.14645, 39.…
#> $ WaldStatistic_treatment_folic_acid_vs_control <dbl> 29.930599, 18.716176, 1…
#> $ WaldPvalue_Intercept                          <dbl> 6.940570e-159, 0.000000…
#> $ WaldPvalue_treatment_folic_acid_vs_control    <dbl> 7.870297e-197, 3.654651…
#> $ betaConv                                      <lgl> TRUE, TRUE, TRUE, TRUE,…
#> $ betaIter                                      <dbl> 2, 3, 3, 2, 3, 4, 3, 2,…
#> $ deviance                                      <dbl> 81.60880, 115.99167, 84…
#> $ maxCooks                                      <dbl> 0.54053191, 0.36317770,…
#> $ control_rep1                                  <dbl> 7.813753e+01, 2.313604e…
#> $ control_rep2                                  <dbl> 46.84188, 2998.40658, 3…
#> $ control_rep3                                  <dbl> 5.622846e+01, 2.733252e…
#> $ fa_day7_rep1                                  <dbl> 25465.2621, 182255.3222…
#> $ fa_day7_rep2                                  <dbl> 25909.906, 121258.268, …
#> $ fa_day7_rep3                                  <dbl> 23817.0835, 104478.2110…