Trimmed Mean of M-Values (TMM) Normalization

TMM normalization is recommended for RNA-seq data generally when the majority of genes are not differentially expressed. We use this as a quality control tool when plotting counts per gene.

tmm(object)

# S4 method for bcbioRNADataSet
tmm(object)

# S4 method for DESeqDataSet
tmm(object)

# S4 method for matrix
tmm(object)

Arguments

object
Object.

Value

matrix.

Examples

data(bcb, dds)

# bcbioRNADataSet
tmm(bcb) %>% head

#>                      group1_1   group1_2   group2_1   group2_2
#> ENSMUSG00000002459   28.63256    8.86970   37.16232   18.59052
#> ENSMUSG00000004768  985.69787  792.49940  848.79742  497.79582
#> ENSMUSG00000005886 4223.48470 3975.93096 4899.91440 2662.40847
#> ENSMUSG00000016918 1670.70126 1424.08740  959.50781  582.67226
#> ENSMUSG00000025774    2.46352    0.00000    0.00000    0.00000
#> ENSMUSG00000025776  145.77000   80.46812   84.60875  105.05135

# NOT RUN {
# DESeqDataSet
tmm(dds) %>% head

# matrix
assay(bcb) %>% tmm %>% head
# }